chore(0.11.11cd): 96c-T5 — KEEP 28 class relocate 進 frontend_contracts/（純位移零行為變更）
CD-96c-7 純機械搬移（比照 feature/87 db 模組化）：28 個 KEEP-justified pytest class
從單體 test_frontend_lint.py 搬進新套件 tests/unit/frontend_contracts/ 五分檔：
- test_contract_api_routes.py（7 class）：前端 fetch URL↔後端 router 路徑契約
- test_contract_lifecycle.py（3）：interval/listener/savedState 資源生命週期
- test_contract_animation.py（6）：ghost-fly/race/timeline 競態契約
- test_contract_desktop.py（5）：PyWebView/桌面 gating + Python-context 注入
- test_contract_layout.py（7）：CSS specificity/breakpoint 單一真理/Jinja-produced-class/資產契約

module-level 路徑常數依 AST 傳遞閉包**複製非剪走**（CD-96c-7）——源檔殘留 class 仍引用同名常數，
故源檔全數保留（source 收集無 NameError）。多重定義常數取首個（相同）定義避 F811。

唯一必要調整（非行為/斷言變更）：新套件深一層目錄，__file__-相對 repo-root 錨統一 +1 級
（Path(__file__).parent×3→×4 共 65 處、parents[2]→parents[3] 共 3 處）；同一 repo root、
零 assert/邏輯改。header import 各檔按實際需求裁（desktop 的 json/re 未用故不 copy）。

純位移證明：collect-only Class::method set 180==180 identical（diff 空）；28/28 class body
byte-level 與原檔相同（反轉路徑調整後）；frontend_contracts 180 passed；source 580 passed
（無 NameError）；source+new = 760 = 原 760；source git diff = 0 added / 2966 deleted（純刪）。

殘留：3 個 ruff nit（F841 m / F541 f-string / F401 local re）為源檔既存 debt 的 verbatim
搬移，改動會破壞 pure-move 保證；tests/ 在 ruff exclude 內 CI 不掃。
