+++ tests/unit/frontend_contracts/test_contract_lifecycle.py
@@ -1,0 +1,172 @@
+"""前端契約守衛（KEEP，跨檔 contract）— 由 test_frontend_lint.py 拆出（96c T5，純搬移零行為變更）。
+
+module-level 路徑常數為源檔複製（CD-96c-7：源檔殘留 class 仍引用同名常數，故複製非剪走）。
+"""
+from pathlib import Path
+
+BATCH_JS = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "state" / "batch.js"
+SEARCH_FLOW_JS = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "state" / "search-flow.js"
+BASE_JS = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "state" / "base.js"
+SETTINGS_CONFIG_JS    = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "settings" / "state-config.js"
+SETTINGS_PROVIDERS_JS = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "settings" / "state-providers.js"
+MAIN_JS = Path(__file__).parent.parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "main.js"
+
+
+class TestBatchIntervalGuard:
+    """T1(40b): 守衛 batch/translate checkInterval 具名 ref + cleanupForNavigation 明確清理"""
+
+    def _batch(self):
+        return BATCH_JS.read_text(encoding="utf-8")
+
+    def _search_flow(self):
+        return SEARCH_FLOW_JS.read_text(encoding="utf-8")
+
+    def _base(self):
+        return BASE_JS.read_text(encoding="utf-8")
+
+    def test_batch_check_interval_assigned(self):
+        """batch.js searchAll() 使用 this._batchCheckInterval = setInterval"""
+        js = self._batch()
+        assert "this._batchCheckInterval = setInterval" in js, \
+            "batch.js searchAll() 應將 setInterval 賦值給 this._batchCheckInterval（具名 ref）"
+
+    def test_translate_check_interval_assigned(self):
+        """batch.js translateAll() 使用 this._translateCheckInterval = setInterval"""
+        js = self._batch()
+        assert "this._translateCheckInterval = setInterval" in js, \
+            "batch.js translateAll() 應將 setInterval 賦值給 this._translateCheckInterval（具名 ref）"
+
+    def test_batch_interval_self_clear(self):
+        """batch.js searchAll() 內 clearInterval(this._batchCheckInterval) 自清"""
+        js = self._batch()
+        assert "clearInterval(this._batchCheckInterval)" in js, \
+            "batch.js searchAll() 應在條件成立時 clearInterval(this._batchCheckInterval)"
+
+    def test_translate_interval_self_clear(self):
+        """batch.js translateAll() 內 clearInterval(this._translateCheckInterval) 自清"""
+        js = self._batch()
+        assert "clearInterval(this._translateCheckInterval)" in js, \
+            "batch.js translateAll() 應在條件成立時 clearInterval(this._translateCheckInterval)"
+
+    def test_cleanup_clears_batch_interval(self):
+        """search-flow.js cleanupForNavigation() 明確清除 this._batchCheckInterval"""
+        js = self._search_flow()
+        assert "clearInterval(this._batchCheckInterval)" in js, \
+            "search-flow.js cleanupForNavigation() 應明確 clearInterval(this._batchCheckInterval)"
+
+    def test_cleanup_clears_translate_interval(self):
+        """search-flow.js cleanupForNavigation() 明確清除 this._translateCheckInterval"""
+        js = self._search_flow()
+        assert "clearInterval(this._translateCheckInterval)" in js, \
+            "search-flow.js cleanupForNavigation() 應明確 clearInterval(this._translateCheckInterval)"
+
+    def test_base_declares_batch_check_interval(self):
+        """base.js state 初始化包含 _batchCheckInterval 欄位"""
+        js = self._base()
+        assert "_batchCheckInterval" in js, \
+            "base.js 應在初始 state 宣告 _batchCheckInterval: null"
+
+    def test_base_declares_translate_check_interval(self):
+        """base.js state 初始化包含 _translateCheckInterval 欄位"""
+        js = self._base()
+        assert "_translateCheckInterval" in js, \
+            "base.js 應在初始 state 宣告 _translateCheckInterval: null"
+
+
+class TestTimerListenerGuard:
+    """T2(40b): 守衛 main.js window listener 具名 ref + cleanup removeEventListener"""
+
+    def _index(self):
+        return MAIN_JS.read_text(encoding="utf-8")
+
+    def _base(self):
+        return BASE_JS.read_text(encoding="utf-8")
+
+    def test_index_uses_set_timer_for_cover_height(self):
+        """main.js $watch('searchResults') 使用 _setTimer('updateCoverHeight'"""
+        js = self._index()
+        assert "_setTimer('updateCoverHeight'" in js, \
+            "main.js $watch('searchResults') 應改用 _setTimer('updateCoverHeight', ...) 取代裸 setTimeout"
+
+    def test_index_no_bare_settimeout_for_cover_height(self):
+        """main.js 不含裸 setTimeout(() => this._updateCoverHeight()"""
+        js = self._index()
+        assert "setTimeout(() => this._updateCoverHeight()" not in js, \
+            "main.js 仍含裸 setTimeout(() => this._updateCoverHeight()，應改為 _setTimer"
+
+    def test_index_pywebview_handler_assigned(self):
+        """main.js pywebview-files listener 賦值給 this._pywebviewFilesHandler"""
+        js = self._index()
+        assert "this._pywebviewFilesHandler =" in js, \
+            "main.js 應將 pywebview-files handler 賦值給 this._pywebviewFilesHandler（具名 ref）"
+
+    def test_index_resize_handler_assigned(self):
+        """main.js resize listener 賦值給 this._resizeHandler"""
+        js = self._index()
+        assert "this._resizeHandler =" in js, \
+            "main.js 應將 resize handler 賦值給 this._resizeHandler（具名 ref）"
+
+    def test_index_cleanup_removes_pywebview_listener(self):
+        """main.js cleanup() 含 removeEventListener('pywebview-files', this._pywebviewFilesHandler)"""
+        js = self._index()
+        assert "removeEventListener('pywebview-files', this._pywebviewFilesHandler)" in js, \
+            "main.js cleanup() 應 removeEventListener('pywebview-files', this._pywebviewFilesHandler)"
+
+    def test_index_cleanup_removes_resize_listener(self):
+        """main.js cleanup() 含 removeEventListener('resize', this._resizeHandler)"""
+        js = self._index()
+        assert "removeEventListener('resize', this._resizeHandler)" in js, \
+            "main.js cleanup() 應 removeEventListener('resize', this._resizeHandler)"
+
+    def test_base_declares_pywebview_handler(self):
+        """base.js state 初始化包含 _pywebviewFilesHandler 欄位"""
+        js = self._base()
+        assert "_pywebviewFilesHandler" in js, \
+            "base.js 應在初始 state 宣告 _pywebviewFilesHandler: null"
+
+    def test_base_declares_resize_handler(self):
+        """base.js state 初始化包含 _resizeHandler 欄位"""
+        js = self._base()
+        assert "_resizeHandler" in js, \
+            "base.js 應在初始 state 宣告 _resizeHandler: null"
+
+
+class TestAutoFetchDirtyStateGuard:
+    """39a-PR-fix P2: 守衛 auto-fallback 後同步 savedState，防止誤觸 dirty state"""
+
+    def _js(self):
+        return SETTINGS_PROVIDERS_JS.read_text(encoding="utf-8")
+
+    def _config_js(self):
+        return SETTINGS_CONFIG_JS.read_text(encoding="utf-8")
+
+    def test_gemini_fallback_syncs_saved_state(self):
+        """testGeminiConnection() auto-fallback 後同步 savedState.geminiModel"""
+        js = self._js()
+        assert "this.savedState.geminiModel" in js, \
+            "settings.js testGeminiConnection() auto-fallback 後應同步 this.savedState.geminiModel，否則 isDirty 誤判"
+
+    def test_openai_fallback_syncs_saved_state(self):
+        """fetchOpenAIModels() auto-assign 後同步 savedState.openaiModel"""
+        js = self._js()
+        assert "this.savedState.openaiModel" in js, \
+            "settings.js fetchOpenAIModels() auto-assign 後應同步 this.savedState.openaiModel，否則 isDirty 誤判"
+
+    def test_openai_config_saves_use_custom_model(self):
+        """saveConfig() openai 區段應含 use_custom_model，以便重載後還原 custom/select 模式"""
+        js = self._config_js()
+        assert "use_custom_model: this.openaiUseCustomModel" in js, \
+            "settings/state-config.js saveConfig() 的 openai 物件應含 use_custom_model: this.openaiUseCustomModel，否則重載後 custom 模式丟失"
+
+    def test_openai_config_loads_use_custom_model(self):
+        """loadConfig() 應從 config 還原 openaiUseCustomModel，而非固定從 false 重設"""
+        js = self._config_js()
+        assert "config.translate.openai?.use_custom_model" in js, \
+            "settings/state-config.js loadConfig() 應含 config.translate.openai?.use_custom_model 讀取，否則重載後 custom 模式無法還原"
+
+    def test_fetch_openai_models_has_source_param(self):
+        """fetchOpenAIModels() 應接受 source 參數，區分 auto-fetch 與手動 Fetch"""
+        js = self._js()
+        assert "source = 'manual'" in js, \
+            "settings.js fetchOpenAIModels() 應含 source = 'manual' 預設參數，避免共享 boolean 競態"
+
