+++ tests/unit/test_frontend_lint.py
@@ -37,64 +37,6 @@ THEME_COLOR_DIM = "#2a303c"
 THEME_COLOR_LIGHT = "#ffffff"
 
 
-class TestAppleTouchIconThemeColor:
-    """TASK-81a-T8 (US-4): apple-touch-icon link + 動態 theme-color meta 守衛。
-
-    - head 有 apple-touch-icon link 指向 /static/apple-touch-icon.png
-    - head 有 theme-color meta，content 為 Jinja 條件，兩白名單 hex 都在
-    - x-init 的 $watch('theme') 會更新 meta[name=theme-color] content（dim/light 雙向）
-    - apple-touch-icon.png 存在且為 180×180（PIL）
-    """
-
-    def _base(self):
-        return BASE_HTML_T76.read_text(encoding="utf-8")
-
-    def test_head_has_apple_touch_icon(self):
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._base(), "html.parser")
-        link = soup.find("link", rel="apple-touch-icon")
-        assert link is not None, "base.html head 缺 <link rel=\"apple-touch-icon\">"
-        assert link.get("href") == "/static/apple-touch-icon.png", \
-            f"apple-touch-icon href 應為 /static/apple-touch-icon.png，實為 {link.get('href')!r}"
-
-    def test_head_has_theme_color_meta(self):
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._base(), "html.parser")
-        meta = soup.find("meta", attrs={"name": "theme-color"})
-        assert meta is not None, "base.html head 缺 <meta name=\"theme-color\">"
-        content = meta.get("content", "")
-        # 初值為 Jinja 條件：{{ '#2a303c' if theme == 'dim' else '#ffffff' }}
-        assert "theme == 'dim'" in content or "theme=='dim'" in content, \
-            f"theme-color content 應為隨 theme 的 Jinja 條件，實為 {content!r}"
-        assert THEME_COLOR_DIM in content and THEME_COLOR_LIGHT in content, \
-            f"theme-color content 須含 dim/light 兩白名單 hex，實為 {content!r}"
-
-    def test_theme_watch_updates_theme_color(self):
-        """$watch('theme', ...) 區塊會把 dim/light hex 寫進 meta[name=theme-color].content"""
-        html = self._base()
-        # 找出更新 theme-color 的 $watch callback（容忍 &quot; HTML escape）
-        pattern = re.compile(
-            r"\$watch\(\s*['\"]theme['\"].*?theme-color.*?\.content.*?"
-            + re.escape(THEME_COLOR_DIM) + r".*?" + re.escape(THEME_COLOR_LIGHT),
-            re.DOTALL,
-        )
-        assert pattern.search(html), (
-            "base.html x-init 缺 $watch('theme') 更新 meta[name=theme-color].content "
-            f"為 {THEME_COLOR_DIM}/{THEME_COLOR_LIGHT} 的區塊（動態狀態列色被移除？）"
-        )
-
-    def test_apple_touch_icon_png_exists_and_180(self):
-        from PIL import Image
-        assert APPLE_TOUCH_ICON_PNG.exists(), \
-            f"apple-touch-icon.png 不存在：{APPLE_TOUCH_ICON_PNG}（跑 tools/gen_apple_touch_icon.py）"
-        with Image.open(APPLE_TOUCH_ICON_PNG) as img:
-            assert img.size == (180, 180), \
-                f"apple-touch-icon.png 應為 180×180，實為 {img.size}"
-            # apple-touch-icon 須不透明品牌底（iOS 自動加圓角，透明背景會變黑/白底不一致）
-            assert img.mode == "RGB", \
-                f"apple-touch-icon.png 須為 RGB 不透明（無 alpha），實為 {img.mode}"
-            assert img.getpixel((0, 0)) == (26, 26, 46), \
-                f"apple-touch-icon.png 四角須為品牌底 #1a1a2e，實為 {img.getpixel((0, 0))}"
 
 
 # [lint-guard: migrate→96e] CSS 半邊已遷 css-guard CG-XP-04（theme.css view-transition anchors）；
@@ -396,126 +338,11 @@ class TestReadonlyConfirmGuard:
             assert "風味" not in text, f"scanner.readonly_confirm_modal.{key} 不得含「風味」二字: {text!r}"
 
 
-class TestBatchIntervalGuard:
-    """T1(40b): 守衛 batch/translate checkInterval 具名 ref + cleanupForNavigation 明確清理"""
-
-    def _batch(self):
-        return BATCH_JS.read_text(encoding="utf-8")
-
-    def _search_flow(self):
-        return SEARCH_FLOW_JS.read_text(encoding="utf-8")
-
-    def _base(self):
-        return BASE_JS.read_text(encoding="utf-8")
-
-    def test_batch_check_interval_assigned(self):
-        """batch.js searchAll() 使用 this._batchCheckInterval = setInterval"""
-        js = self._batch()
-        assert "this._batchCheckInterval = setInterval" in js, \
-            "batch.js searchAll() 應將 setInterval 賦值給 this._batchCheckInterval（具名 ref）"
-
-    def test_translate_check_interval_assigned(self):
-        """batch.js translateAll() 使用 this._translateCheckInterval = setInterval"""
-        js = self._batch()
-        assert "this._translateCheckInterval = setInterval" in js, \
-            "batch.js translateAll() 應將 setInterval 賦值給 this._translateCheckInterval（具名 ref）"
-
-    def test_batch_interval_self_clear(self):
-        """batch.js searchAll() 內 clearInterval(this._batchCheckInterval) 自清"""
-        js = self._batch()
-        assert "clearInterval(this._batchCheckInterval)" in js, \
-            "batch.js searchAll() 應在條件成立時 clearInterval(this._batchCheckInterval)"
-
-    def test_translate_interval_self_clear(self):
-        """batch.js translateAll() 內 clearInterval(this._translateCheckInterval) 自清"""
-        js = self._batch()
-        assert "clearInterval(this._translateCheckInterval)" in js, \
-            "batch.js translateAll() 應在條件成立時 clearInterval(this._translateCheckInterval)"
-
-    def test_cleanup_clears_batch_interval(self):
-        """search-flow.js cleanupForNavigation() 明確清除 this._batchCheckInterval"""
-        js = self._search_flow()
-        assert "clearInterval(this._batchCheckInterval)" in js, \
-            "search-flow.js cleanupForNavigation() 應明確 clearInterval(this._batchCheckInterval)"
-
-    def test_cleanup_clears_translate_interval(self):
-        """search-flow.js cleanupForNavigation() 明確清除 this._translateCheckInterval"""
-        js = self._search_flow()
-        assert "clearInterval(this._translateCheckInterval)" in js, \
-            "search-flow.js cleanupForNavigation() 應明確 clearInterval(this._translateCheckInterval)"
-
-    def test_base_declares_batch_check_interval(self):
-        """base.js state 初始化包含 _batchCheckInterval 欄位"""
-        js = self._base()
-        assert "_batchCheckInterval" in js, \
-            "base.js 應在初始 state 宣告 _batchCheckInterval: null"
-
-    def test_base_declares_translate_check_interval(self):
-        """base.js state 初始化包含 _translateCheckInterval 欄位"""
-        js = self._base()
-        assert "_translateCheckInterval" in js, \
-            "base.js 應在初始 state 宣告 _translateCheckInterval: null"
 
 
 MAIN_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "main.js"
 
 
-class TestTimerListenerGuard:
-    """T2(40b): 守衛 main.js window listener 具名 ref + cleanup removeEventListener"""
-
-    def _index(self):
-        return MAIN_JS.read_text(encoding="utf-8")
-
-    def _base(self):
-        return BASE_JS.read_text(encoding="utf-8")
-
-    def test_index_uses_set_timer_for_cover_height(self):
-        """main.js $watch('searchResults') 使用 _setTimer('updateCoverHeight'"""
-        js = self._index()
-        assert "_setTimer('updateCoverHeight'" in js, \
-            "main.js $watch('searchResults') 應改用 _setTimer('updateCoverHeight', ...) 取代裸 setTimeout"
-
-    def test_index_no_bare_settimeout_for_cover_height(self):
-        """main.js 不含裸 setTimeout(() => this._updateCoverHeight()"""
-        js = self._index()
-        assert "setTimeout(() => this._updateCoverHeight()" not in js, \
-            "main.js 仍含裸 setTimeout(() => this._updateCoverHeight()，應改為 _setTimer"
-
-    def test_index_pywebview_handler_assigned(self):
-        """main.js pywebview-files listener 賦值給 this._pywebviewFilesHandler"""
-        js = self._index()
-        assert "this._pywebviewFilesHandler =" in js, \
-            "main.js 應將 pywebview-files handler 賦值給 this._pywebviewFilesHandler（具名 ref）"
-
-    def test_index_resize_handler_assigned(self):
-        """main.js resize listener 賦值給 this._resizeHandler"""
-        js = self._index()
-        assert "this._resizeHandler =" in js, \
-            "main.js 應將 resize handler 賦值給 this._resizeHandler（具名 ref）"
-
-    def test_index_cleanup_removes_pywebview_listener(self):
-        """main.js cleanup() 含 removeEventListener('pywebview-files', this._pywebviewFilesHandler)"""
-        js = self._index()
-        assert "removeEventListener('pywebview-files', this._pywebviewFilesHandler)" in js, \
-            "main.js cleanup() 應 removeEventListener('pywebview-files', this._pywebviewFilesHandler)"
-
-    def test_index_cleanup_removes_resize_listener(self):
-        """main.js cleanup() 含 removeEventListener('resize', this._resizeHandler)"""
-        js = self._index()
-        assert "removeEventListener('resize', this._resizeHandler)" in js, \
-            "main.js cleanup() 應 removeEventListener('resize', this._resizeHandler)"
-
-    def test_base_declares_pywebview_handler(self):
-        """base.js state 初始化包含 _pywebviewFilesHandler 欄位"""
-        js = self._base()
-        assert "_pywebviewFilesHandler" in js, \
-            "base.js 應在初始 state 宣告 _pywebviewFilesHandler: null"
-
-    def test_base_declares_resize_handler(self):
-        """base.js state 初始化包含 _resizeHandler 欄位"""
-        js = self._base()
-        assert "_resizeHandler" in js, \
-            "base.js 應在初始 state 宣告 _resizeHandler: null"
 
 
 class TestSettingsCleanupBypassGuard:
@@ -571,142 +398,6 @@ class TestSettingsCleanupBypassGuard:
 LOCALES_ROOT = Path(__file__).parent.parent.parent / "locales"
 
 
-class TestJellyfinCheckManualGuard:
-    """40c-T2: 守衛 Jellyfin check 改為手動觸發的所有前端不變式"""
-
-    def _html(self):
-        return SCANNER_HTML.read_text(encoding="utf-8")
-
-    def _js(self):
-        return SCANNER_SCAN_JS.read_text(encoding="utf-8")
-
-    def test_no_auto_trigger_in_init(self):
-        """init() 後的 loadStats 呼叫後，不應緊接 checkJellyfinImages()"""
-        # 確認 checkJellyfinImages() 只透過 @click 觸發，不在 init() 或 loadStats 後出現
-        js = self._js()
-        assert "this.loadStats();\n        this.checkJellyfinImages();" not in js, \
-            "scanner.js init() 仍含自動觸發 checkJellyfinImages()"
-
-    def test_jellyfin_check_state_declared(self):
-        """Alpine data 宣告 jellyfinCheckState 欄位"""
-        js = self._js()
-        assert "jellyfinCheckState: 'idle'" in js, \
-            "scanner.js 缺少 jellyfinCheckState: 'idle' 初始化宣告"
-
-    def test_jellyfin_check_controller_declared(self):
-        """Alpine data 宣告 _jellyfinCheckController 欄位"""
-        js = self._js()
-        assert "_jellyfinCheckController: null" in js, \
-            "scanner.js 缺少 _jellyfinCheckController: null 初始化宣告"
-
-    def test_abort_controller_used_in_check(self):
-        """checkJellyfinImages() 建立 AbortController"""
-        js = self._js()
-        assert "new AbortController()" in js, \
-            "scanner.js checkJellyfinImages() 缺少 new AbortController()"
-
-    def test_abort_called_in_cleanup(self):
-        """cleanup 回呼內含 _jellyfinCheckController.abort()"""
-        js = self._js()
-        assert "_jellyfinCheckController.abort()" in js, \
-            "scanner.js cleanup 缺少 _jellyfinCheckController.abort()"
-
-    def test_jellyfin_check_state_reset_in_cleanup(self):
-        """cleanup 回呼補上 jellyfinCheckState = 'idle' 重設"""
-        js = self._js()
-        assert "jellyfinCheckState = 'idle'" in js, \
-            "scanner.js cleanup 缺少 jellyfinCheckState = 'idle' 重設"
-
-    def test_should_warn_checks_jellyfin_checking(self):
-        """shouldWarnBeforeLeave() 含 jellyfinCheckState === 'checking' 判斷"""
-        js = self._js()
-        assert "jellyfinCheckState === 'checking'" in js, \
-            "scanner.js shouldWarnBeforeLeave() 缺少 jellyfinCheckState === 'checking' 判斷"
-
-    def test_trigger_button_click_handler(self):
-        """觸發按鈕 @click 呼叫 checkJellyfinImages()"""
-        html = self._html()
-        assert '@click="checkJellyfinImages()"' in html, \
-            "scanner.html 缺少 @click=\"checkJellyfinImages()\" 觸發按鈕"
-
-    def test_no_auto_trigger_after_generate(self):
-        """generate SSE done 事件後無自動呼叫 checkJellyfinImages()"""
-        js = self._js()
-        # loadStats 後面不應接 checkJellyfinImages（generate 路徑）
-        assert "this.loadStats();\n                    this.checkJellyfinImages();" not in js, \
-            "scanner.js generate done 路徑仍自動呼叫 checkJellyfinImages()"
-
-    def test_clear_cache_resets_jellyfin_state(self):
-        """clearCache 成功後含 jellyfinCheckState = 'idle' 重設"""
-        js = self._js()
-        # jellyfinImageVisible = false 後緊接 jellyfinCheckState = 'idle'
-        assert "jellyfinImageVisible = false" in js, \
-            "scanner.js clearCache 缺少 jellyfinImageVisible = false"
-        # jellyfinCheckState = 'idle' 在 clearCache 函數中也必須出現
-        # （在 shouldWarnBeforeLeave 和 beforeLeave 都有，此守衛確認 clearCache 路徑有）
-        # 用計數確認至少 2 處（cleanup + clearCache，shouldWarnBeforeLeave 判斷不算 assignment）
-        count = js.count("jellyfinCheckState = 'idle'")
-        assert count >= 2, \
-            f"scanner.js jellyfinCheckState = 'idle' 出現 {count} 次，期望 >= 2（cleanup + clearCache）"
-
-    def test_trigger_row_xshow_uses_jellyfin_image_visible(self):
-        """T3(40c) / T-d4 / 72d-codexP2: 觸發列 x-show 用正向白名單 gate（fail-closed，含 kodi）"""
-        from bs4 import BeautifulSoup
-        html = self._html()
-        soup = BeautifulSoup(html, "html.parser")
-        # There are multiple nfo-update-row divs; the jellyfin trigger row is the one
-        # whose x-show references jellyfinImageVisible (not nfoUpdateVisible etc.)
-        rows = soup.find_all("div", class_="nfo-update-row")
-        jellyfin_row = next(
-            (r for r in rows if "jellyfinImageVisible" in r.get("x-show", "") and "config" in r.get("x-show", "")),
-            None
-        )
-        assert jellyfin_row is not None, \
-            "scanner.html 找不到含 jellyfinImageVisible + config 的 nfo-update-row element"
-        xshow = jellyfin_row.get("x-show", "")
-        # 正向白名單（fail-closed）：config={} / undefined 時 gate 為 false，不顯示
-        assert "['jellyfin', 'emby', 'kodi'].includes(config?.scraper?.external_manager)" in xshow, \
-            f"nfo-update-row x-show 應使用正向白名單 .includes() gate（fail-closed），實際: {xshow!r}"
-        assert "!jellyfinImageVisible" in xshow, \
-            f"nfo-update-row x-show 應含 !jellyfinImageVisible，實際: {xshow!r}"
-        # forbidden：舊 jellyfin_emby gate 與 interim !='off'（fail-open）皆不得殘留
-        assert "=== 'jellyfin_emby'" not in html, \
-            "scanner.html 仍殘留舊 === 'jellyfin_emby' gate"
-        assert "config?.scraper?.external_manager !== 'off' && !jellyfinImageVisible" not in html, \
-            "scanner.html 觸發列仍用 interim !== 'off' gate（undefined 會 fail-open，須改正向白名單）"
-        assert "config?.scraper?.jellyfin_mode && !jellyfinImageVisible" not in html, \
-            "scanner.html 觸發列 x-show 仍使用舊的 jellyfin_mode 讀取點（應已 repoint 為 external_manager）"
-
-    def test_check_jellyfin_method_gate_is_fail_closed(self):
-        """72d-codexP2: checkJellyfinImages() 方法端 gate 用正向白名單，config 未載入時 fail-closed 不打 API"""
-        js = self._js()
-        assert "async checkJellyfinImages()" in js, "state-scan.js 找不到 checkJellyfinImages() 方法"
-        # 正向白名單 early-return（fail-closed）；此字串為該 gate 獨有
-        assert "!['jellyfin', 'emby', 'kodi'].includes(this.config?.scraper?.external_manager)" in js, \
-            "checkJellyfinImages() 應以正向白名單 early-return（fail-closed），不可用 === 'off'（undefined fail-open）"
-        # forbidden：舊 fail-open gate（undefined === 'off' 為 false → 不 return → 打 /jellyfin-check）
-        assert "this.config?.scraper?.external_manager === 'off'" not in js, \
-            "state-scan.js 仍殘留 external_manager === 'off' gate（undefined 會 fail-open）"
-
-    def test_trigger_row_done_state_text_present(self):
-        """T3(40c) Codex fix: 觸發列包含 done 狀態顯示文字"""
-        html = self._html()
-        assert "jellyfinCheckState === 'done'" in html, \
-            "scanner.html 觸發列缺少 done 狀態文字顯示"
-        assert "jellyfin_check_done_ok" in html, \
-            "scanner.html 觸發列缺少 jellyfin_check_done_ok i18n key 引用"
-
-    def test_jellyfin_update_done_resets_check_state(self):
-        """T3(40c) Codex fix: jellyfin-update done handler 重設 jellyfinCheckState = 'idle'"""
-        js = self._js()
-        # 確認 runJellyfinImageUpdate 的 done 分支有三個重設欄位
-        assert "this.jellyfinImageVisible = false" in js, \
-            "scanner.js runJellyfinImageUpdate done 缺少 jellyfinImageVisible = false 重設"
-        assert "this.jellyfinImageCount = 0" in js, \
-            "scanner.js runJellyfinImageUpdate done 缺少 jellyfinImageCount = 0 重設"
-        # jellyfinCheckState = 'idle' 重設（update done 分支使用 this. 前綴）
-        assert "this.jellyfinCheckState = 'idle'" in js, \
-            "scanner.js runJellyfinImageUpdate done 缺少 this.jellyfinCheckState = 'idle' 重設"
 
 
 class TestShowcaseKeyboardGuard:
@@ -999,81 +690,8 @@ NAVIGATION_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / 
 ANIMATIONS_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "animations.js"
 
 
-class TestCodexFixes:
-    """39a Codex review 修正守衛"""
-
-    def _navigation_js(self):
-        return NAVIGATION_JS.read_text(encoding="utf-8")
-
-    def _settings_js(self):
-        return SETTINGS_PROVIDERS_JS.read_text(encoding="utf-8")
-
-    def test_loadmore_no_currentindex_assignment(self):
-        """F1：loadMore() 成功分支不含 this.currentIndex = 賦值"""
-        js = self._navigation_js()
-        # 找到 loadMore 函數體，截取到 finally 區塊結束
-        start = js.find("async loadMore(trigger")
-        assert start != -1, "navigation.js 找不到 async loadMore(trigger ...) 函數"
-        # 截取 loadMore 函數體（到函數結尾）
-        func_body = js[start:]
-        # 找到 finally { ... } 後的第一個右大括號（函數結束）
-        finally_pos = func_body.find("finally {")
-        if finally_pos != -1:
-            # 截取 loadMore 函數範圍：從函數開始到 finally 區塊後的 } 結尾
-            end_pos = func_body.find("}", finally_pos + len("finally {"))
-            # 再找外層函數的 }
-            end_pos = func_body.find("},", end_pos + 1)
-            func_body = func_body[:end_pos] if end_pos != -1 else func_body
-        # 確認函數體內不含 this.currentIndex = 賦值（有空格的賦值語句）
-        assert "this.currentIndex =" not in func_body, \
-            "navigation.js loadMore() 成功分支不應含 this.currentIndex = 賦值（破壞 shared state contract）"
-
-    def test_gemini_model_fallback_includes_check(self):
-        """F2：testGeminiConnection() 成功後包含 includes() 檢查舊 model 是否在 allowlist"""
-        js = self._settings_js()
-        assert "modelNames.includes(this.form.geminiModel)" in js or \
-               "includes(this.form.geminiModel)" in js, \
-            "settings.js testGeminiConnection() 成功後應含 includes(this.form.geminiModel) allowlist 檢查"
 
 
-class TestAutoFetchDirtyStateGuard:
-    """39a-PR-fix P2: 守衛 auto-fallback 後同步 savedState，防止誤觸 dirty state"""
-
-    def _js(self):
-        return SETTINGS_PROVIDERS_JS.read_text(encoding="utf-8")
-
-    def _config_js(self):
-        return SETTINGS_CONFIG_JS.read_text(encoding="utf-8")
-
-    def test_gemini_fallback_syncs_saved_state(self):
-        """testGeminiConnection() auto-fallback 後同步 savedState.geminiModel"""
-        js = self._js()
-        assert "this.savedState.geminiModel" in js, \
-            "settings.js testGeminiConnection() auto-fallback 後應同步 this.savedState.geminiModel，否則 isDirty 誤判"
-
-    def test_openai_fallback_syncs_saved_state(self):
-        """fetchOpenAIModels() auto-assign 後同步 savedState.openaiModel"""
-        js = self._js()
-        assert "this.savedState.openaiModel" in js, \
-            "settings.js fetchOpenAIModels() auto-assign 後應同步 this.savedState.openaiModel，否則 isDirty 誤判"
-
-    def test_openai_config_saves_use_custom_model(self):
-        """saveConfig() openai 區段應含 use_custom_model，以便重載後還原 custom/select 模式"""
-        js = self._config_js()
-        assert "use_custom_model: this.openaiUseCustomModel" in js, \
-            "settings/state-config.js saveConfig() 的 openai 物件應含 use_custom_model: this.openaiUseCustomModel，否則重載後 custom 模式丟失"
-
-    def test_openai_config_loads_use_custom_model(self):
-        """loadConfig() 應從 config 還原 openaiUseCustomModel，而非固定從 false 重設"""
-        js = self._config_js()
-        assert "config.translate.openai?.use_custom_model" in js, \
-            "settings/state-config.js loadConfig() 應含 config.translate.openai?.use_custom_model 讀取，否則重載後 custom 模式無法還原"
-
-    def test_fetch_openai_models_has_source_param(self):
-        """fetchOpenAIModels() 應接受 source 參數，區分 auto-fetch 與手動 Fetch"""
-        js = self._js()
-        assert "source = 'manual'" in js, \
-            "settings.js fetchOpenAIModels() 應含 source = 'manual' 預設參數，避免共享 boolean 競態"
 
 
 MOTION_LAB_STATE_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "motion-lab-state.js"
@@ -1433,78 +1051,6 @@ PATH_UTILS_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / 
 FILE_LIST_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "search" / "state" / "file-list.js"
 
 
-class TestUserTagsApiGuard:
-    """41b-T3: 確保 confirmAddTag 和 removeUserTag 改接 /api/user-tags API（method folded）"""
-
-    def _result_card(self):
-        return RESULT_CARD_JS.read_text(encoding="utf-8")
-
-    def _path_utils(self):
-        return PATH_UTILS_JS.read_text(encoding="utf-8")
-
-    def _html(self):
-        return SEARCH_HTML.read_text(encoding="utf-8")
-
-    def _locale(self, name):
-        return json.loads((LOCALES_ROOT / name).read_text(encoding="utf-8"))
-
-    def _get_nested(self, d, dotted_key):
-        keys = dotted_key.split(".")
-        cur = d
-        for k in keys:
-            if not isinstance(cur, dict) or k not in cur:
-                return None
-            cur = cur[k]
-        return cur
-
-    def test_result_card_js_contains(self):
-        """result-card.js 含 API 呼叫、async functions、file-level user_tags、fetch helper"""
-        content = self._result_card()
-        for expected in [
-            "user-tags",
-            "async confirmAddTag()",
-            "async removeUserTag(",
-            "fileList[this.currentFileIndex].user_tags",
-            "currentUserTags()",
-            "fetchUserTagsForCurrent",
-        ]:
-            assert expected in content, f"result-card.js missing: {expected!r}"
-        # not-in guards
-        for forbidden in ["pathToFileUri", "c.user_tags.push(tag)"]:
-            assert forbidden not in content, f"result-card.js should not contain: {forbidden!r}"
-        # fetchUserTagsForCurrent writes to file-level
-        idx = content.find("async fetchUserTagsForCurrent()")
-        assert idx != -1, "result-card.js missing: 'async fetchUserTagsForCurrent()'"
-        func_body = content[idx:idx+800]
-        has_direct = "fileList[this.currentFileIndex].user_tags" in func_body
-        has_captured_ref = ("file.user_tags" in func_body and
-                            "this.fileList?.[this.currentFileIndex]" in func_body)
-        assert has_direct or has_captured_ref, \
-            "fetchUserTagsForCurrent missing file-level user_tags write"
-
-    def test_search_html_contains(self):
-        """search.html 含 user-tags 守衛 + currentUserTags()"""
-        html = self._html()
-        for expected in [
-            "listMode === \'file\'",
-            "fileList[currentFileIndex]?.path",
-            "currentUserTags()",
-        ]:
-            assert expected in html, f"search.html missing: {expected!r}"
-
-    def test_path_utils_and_locales(self):
-        """path-utils.js 無 pathToFileUri + 有 pathToDisplay；locales 含 tag_api_failed"""
-        pu = self._path_utils()
-        assert "pathToFileUri" not in pu, "path-utils.js should not contain: 'pathToFileUri'"
-        assert "pathToDisplay" in pu, "path-utils.js missing: 'pathToDisplay'"
-        # file-list.js user_tags init
-        file_list_content = FILE_LIST_JS.read_text(encoding="utf-8")
-        assert "user_tags: []" in file_list_content, "file-list.js missing: 'user_tags: []'"
-        # locales
-        for locale_file in ["zh_TW.json", "zh_CN.json", "en.json", "ja.json"]:
-            data = self._locale(locale_file)
-            val = self._get_nested(data, "search.error.tag_api_failed")
-            assert val, f"{locale_file} missing: search.error.tag_api_failed key"
 
 class TestShowcaseActressTemplate:
     """Phase 44a-T3: 守衛 showcase.html 含有女優模式 UI 結構（method folded）"""
@@ -1738,32 +1284,6 @@ class TestShowcaseLightboxSentinel:
             f'x-trap.inert 未含 deleteVideoModalOpen（delete modal 開時 trap 未釋放）: {expr!r}'
 
 
-class TestShowcaseAliasGuard:
-    """T5 (45-actress-alias): Frontend Guard — alias injection guard (method folded)"""
-
-    def _js(self):
-        return (
-            SHOWCASE_BASE_JS.read_text(encoding="utf-8") + "\n" +
-            SHOWCASE_ACTRESS_JS.read_text(encoding="utf-8") + "\n" +
-            SHOWCASE_VIDEOS_JS.read_text(encoding="utf-8")
-        )
-
-    def test_alias_js_contains(self):
-        """showcase JS 含 _nameToGroup 宣告 + API + 使用"""
-        js = self._js()
-        for expected in [
-            "var _nameToGroup = {}",
-            "/api/actress-aliases",
-            "_nameToGroup[a.name]",
-            "_nameToGroup[term]",
-        ]:
-            assert expected in js, f"showcase JS missing: {expected!r}"
-        # _checkPreciseActressMatch function body must use _nameToGroup
-        func_start = js.find("async _checkPreciseActressMatch")
-        func_end = js.find("},", func_start)
-        func_body = js[func_start:func_end]
-        assert "_nameToGroup" in func_body, \
-            "showcase JS _checkPreciseActressMatch missing: '_nameToGroup'"
 
 
 # ---------------------------------------------------------------------------
@@ -2014,113 +1534,6 @@ class TestShowcaseAnimationsFluent:
         assert "typeof window.GhostFly?.playLightboxOpen === 'function'" in scope, \
             "search/animations.js missing: typeof guard for playLightboxOpen"
 
-class TestGhostFlyGuards:
-    """T8: Ghost Fly architecture guards (method folded)"""
-
-    def test_ghost_fly_js_and_html_contains(self):
-        """ghost-fly.js exists + loaded in base.html + skipCover support + delegates"""
-        assert Path("web/static/js/shared/ghost-fly.js").exists(), \
-            "web/static/js/shared/ghost-fly.js missing"
-        html = Path("web/templates/base.html").read_text(encoding="utf-8")
-        assert "ghost-fly.js" in html, "base.html missing: 'ghost-fly.js'"
-        ghost_fly_js = Path("web/static/js/shared/ghost-fly.js").read_text(encoding="utf-8")
-        assert "skipCover" in ghost_fly_js, "ghost-fly.js missing: 'skipCover'"
-        for path in [
-            "web/static/js/pages/showcase/animations.js",
-            "web/static/js/pages/search/animations.js",
-        ]:
-            js = Path(path).read_text(encoding="utf-8")
-            assert "GhostFly.playLightboxOpen" in js, f"{path} missing: 'GhostFly.playLightboxOpen'"
-        # search/animations.js fallback
-        search_js = Path("web/static/js/pages/search/animations.js").read_text(encoding="utf-8")
-        lines = search_js.split('\n')
-        ghost_fly_refs = [i for i, line in enumerate(lines) if 'window.GhostFly' in line]
-        assert len(ghost_fly_refs) >= 3, \
-            "search/animations.js missing: at least 3 window.GhostFly references"
-
-    def test_gsap_animating_before_lightbox_open(self):
-        """state-lightbox.js gsap-animating before lightboxOpen = true (ordering)"""
-        content = SHOWCASE_LIGHTBOX_JS.read_text(encoding="utf-8")
-        for fn_name in ("openLightbox(", "openHeroCardLightbox("):
-            idx_fn = content.find(fn_name)
-            assert idx_fn > 0, f"state-lightbox.js missing: {fn_name!r}"
-            fn_scope = content[idx_fn:idx_fn + 4000]
-            idx_animating = fn_scope.find("gsap-animating")
-            idx_open = fn_scope.find("this.lightboxOpen = true")
-            assert idx_animating > 0, f"state-lightbox.js {fn_name} missing: 'gsap-animating'"
-            assert idx_open > 0, f"state-lightbox.js {fn_name} missing: 'lightboxOpen = true'"
-            assert idx_animating < idx_open, \
-                f"state-lightbox.js {fn_name}: gsap-animating must precede lightboxOpen = true"
-
-    # ── 71b-T3: both-restore guard（hide/restore 目標皆為 .lightbox-cover 容器）──
-    # element-bound：regex 抽 OPEN(playGridToLightbox) / CLOSE(playLightboxToGrid)
-    # 各自 function body，斷言兩路 hide + restore 都指向 coverEl 容器（非僅單一 img）。
-    # 非檔案層級的 '.lightbox-cover' 字串存在性檢查（comment 留字串無法騙過）。
-
-    GHOST_FLY_JS = Path("web/static/js/shared/ghost-fly.js")
-
-    def _extract_method_body(self, js, method_name):
-        """抓 `methodName: function (...) {` 物件方法的 body（大括號平衡匹配）。"""
-        pattern = re.compile(
-            re.escape(method_name) + r'\s*:\s*function\s*\([^)]*\)\s*\{',
-            re.DOTALL,
-        )
-        m = pattern.search(js)
-        assert m is not None, f"ghost-fly.js 找不到 {method_name} 方法"
-        start = m.end()  # 位於 { 之後
-        depth = 1
-        i = start
-        while i < len(js) and depth > 0:
-            c = js[i]
-            if c == '{':
-                depth += 1
-            elif c == '}':
-                depth -= 1
-            i += 1
-        return js[start:i - 1]
-
-    def test_open_hide_and_restore_target_is_cover_container(self):
-        """OPEN(playGridToLightbox) body：coverEl 取自 .lightbox-cover 容器，
-        hide（data-ghost-hidden + opacity:0）與 cleanupGhost restore 皆指向 coverEl，
-        而非僅單一 lbImg。"""
-        js = self.GHOST_FLY_JS.read_text(encoding="utf-8")
-        body = self._extract_method_body(js, "playGridToLightbox")
-        # coverEl 由 .lightbox-cover 容器取得（closest / querySelector 任一）
-        assert re.search(r'var\s+coverEl\s*=', body), \
-            "playGridToLightbox body 缺少 coverEl 宣告（應隱 .lightbox-cover 容器而非單一 img）"
-        assert "closest('.lightbox-cover')" in body or "querySelector('.lightbox-cover')" in body, \
-            "playGridToLightbox coverEl 必須取自 .lightbox-cover 容器"
-        # hide 目標是 coverEl（attribute + opacity:0）
-        assert re.search(r"coverEl\.setAttribute\(\s*'data-ghost-hidden'", body), \
-            "playGridToLightbox hide 必須對 coverEl 掛 data-ghost-hidden（容器，非僅 lbImg）"
-        assert re.search(r"gsap\.set\(\s*coverEl\s*,\s*\{\s*opacity:\s*0", body), \
-            "playGridToLightbox hide 必須對 coverEl 設 opacity:0（容器，非僅 lbImg）"
-        # restore 目標是 coverEl（cleanupGhost 帶 coverEl）
-        assert re.search(r"cleanupGhost\(\s*ghost\s*,\s*coverEl", body), \
-            "playGridToLightbox cleanupGhost restore 必須帶 coverEl（容器），非僅 lbImg"
-
-    def test_close_hide_and_restore_target_is_cover_container(self):
-        """CLOSE(playLightboxToGrid) body：coverEl 取自 .lightbox-cover 容器，
-        hide 補掛 data-ghost-hidden + opacity:0，abort 還原 coverEl，
-        normal-complete 的 cleanupGhost restore 參數含 coverEl（與 OPEN 對稱）。"""
-        js = self.GHOST_FLY_JS.read_text(encoding="utf-8")
-        body = self._extract_method_body(js, "playLightboxToGrid")
-        # coverEl 由 .lightbox-cover 容器取得
-        assert re.search(r'var\s+coverEl\s*=', body), \
-            "playLightboxToGrid body 缺少 coverEl 宣告（應隱 .lightbox-cover 容器而非單一 fromImg）"
-        assert "closest('.lightbox-cover')" in body or "querySelector('.lightbox-cover')" in body, \
-            "playLightboxToGrid coverEl 必須取自 .lightbox-cover 容器"
-        # hide 目標是 coverEl（補上 attribute，舊版漏掛）+ opacity:0
-        assert re.search(r"coverEl\.setAttribute\(\s*'data-ghost-hidden'", body), \
-            "playLightboxToGrid hide 必須對 coverEl 掛 data-ghost-hidden（舊版漏掛 → stale-cleanup 兜不到）"
-        assert re.search(r"gsap\.set\(\s*coverEl\s*,\s*\{\s*opacity:\s*0", body), \
-            "playLightboxToGrid hide 必須對 coverEl 設 opacity:0（容器，非僅 fromImg）"
-        # abort 還原 coverEl
-        assert re.search(r"gsap\.set\(\s*coverEl\s*,\s*\{\s*opacity:\s*1", body), \
-            "playLightboxToGrid abort 必須還原 coverEl opacity:1（容器，非僅 fromImg）"
-        # normal-complete restore 含 coverEl（對稱還原來源容器，修舊不對稱）
-        assert re.search(r"cleanupGhost\(\s*ghost\s*,\s*targetImg\s*,\s*coverEl", body), \
-            "playLightboxToGrid cleanupGhost restore 參數必須含 coverEl（對稱還原來源容器，非僅 targetImg）"
 
 
 class TestTutorialSkipPersistsGuard:
@@ -2489,274 +1902,8 @@ SHOWCASE_ANIMATIONS_JS = (
 )
 
 
-class TestModeToggleFadeOutGuard:
-    """T1: 模式切換動畫補 fade-out（playModeCrossfade 4-arg + toggleActressMode callback 延遲翻轉）"""
-
-    def _core_js(self):
-        # toggleActressMode / searchActressFilms → state-actress.js
-        # switchMode → state-videos.js
-        return (
-            SHOWCASE_ACTRESS_JS.read_text(encoding="utf-8") + "\n" +
-            SHOWCASE_VIDEOS_JS.read_text(encoding="utf-8")
-        )
-
-    def _anim_js(self):
-        return SHOWCASE_ANIMATIONS_JS.read_text(encoding="utf-8")
-
-    def _extract_method_body(self, js, method_name):
-        """抓取 Alpine state method（methodName(...) { ... }）函式主體，大括號平衡（容忍 async 前綴）。"""
-        pattern = re.compile(
-            r'(?:^|\n)\s*(?:async\s+)?' + re.escape(method_name) + r'\s*\([^)]*\)\s*\{',
-            re.DOTALL,
-        )
-        m = pattern.search(js)
-        assert m is not None, f"找不到 {method_name} 方法"
-        start = m.end()
-        depth = 1
-        i = start
-        while i < len(js) and depth > 0:
-            c = js[i]
-            if c == '{':
-                depth += 1
-            elif c == '}':
-                depth -= 1
-            i += 1
-        return js[start:i - 1]
-
-    def _extract_property_function_body(self, js, prop_name):
-        """抓取 propName: function (...) { ... } 形式的函式主體，大括號平衡。"""
-        pattern = re.compile(
-            r'\b' + re.escape(prop_name) + r'\s*:\s*function\s*\([^)]*\)\s*\{',
-            re.DOTALL,
-        )
-        m = pattern.search(js)
-        assert m is not None, f"找不到 {prop_name} property function"
-        start = m.end()
-        depth = 1
-        i = start
-        while i < len(js) and depth > 0:
-            c = js[i]
-            if c == '{':
-                depth += 1
-            elif c == '}':
-                depth -= 1
-            i += 1
-        return js[start:i - 1]
-
-    def test_play_mode_crossfade_has_callbacks_param(self):
-        """animations.js playModeCrossfade 簽名包含 4 個參數 (oldMode, newMode, params, callbacks)"""
-        js = self._anim_js()
-        assert re.search(
-            r'playModeCrossfade\s*:\s*function\s*\(\s*oldMode\s*,\s*newMode\s*,\s*params\s*,\s*callbacks\s*\)',
-            js,
-        ), "showcase/animations.js playModeCrossfade 缺少 callbacks 第 4 參數"
-
-    def test_play_mode_crossfade_old_fade_out(self):
-        """playModeCrossfade 函數體含 oldEl fade-out (tl.to(oldEl,...) + clearProps:'opacity')"""
-        js = self._anim_js()
-        body = self._extract_property_function_body(js, 'playModeCrossfade')
-        assert re.search(r'tl\s*\.\s*to\s*\(\s*oldEl', body), \
-            "playModeCrossfade 函數體缺少 oldEl fade-out (tl.to(oldEl,...))"
-        assert re.search(r"clearProps\s*:\s*['\"]opacity['\"]", body), \
-            "playModeCrossfade 函數體缺少 clearProps: 'opacity'（避免 CSS transition 殘留）"
-
-    def test_play_mode_crossfade_new_fade_in_preserved(self):
-        """playModeCrossfade 函數體保留 newEl fade-in（tl.fromTo(newEl,...) + clearProps:'opacity'）"""
-        js = self._anim_js()
-        body = self._extract_property_function_body(js, 'playModeCrossfade')
-        assert re.search(r'(?:tl\s*\.\s*)?fromTo\s*\(\s*newEl', body), \
-            "playModeCrossfade 函數體缺少 newEl fade-in (fromTo(newEl,...))"
-        # newEl 段落（從第一次 newEl 出現到結尾）必須有 clearProps
-        new_idx = body.find('newEl')
-        assert new_idx >= 0, "playModeCrossfade 函數體找不到 newEl 區段"
-        new_section = body[new_idx:]
-        assert re.search(r"clearProps\s*:\s*['\"]opacity['\"]", new_section), \
-            "playModeCrossfade newEl fade-in 段落缺少 clearProps: 'opacity'"
-
-    def test_toggle_actress_mode_uses_callback(self):
-        """toggleActressMode 函數體使用 onOldFadeComplete callback，不直接翻轉 showFavoriteActresses"""
-        js = self._core_js()
-        body = self._extract_method_body(js, 'toggleActressMode')
-        assert 'onOldFadeComplete' in body, \
-            "toggleActressMode 函數體缺少 onOldFadeComplete callback"
-        assert 'playModeCrossfade' in body, \
-            "toggleActressMode 函數體缺少 playModeCrossfade 呼叫"
-        assert not re.search(
-            r'this\.showFavoriteActresses\s*=\s*!\s*this\.showFavoriteActresses',
-            body,
-        ), "toggleActressMode 不應直接翻轉 this.showFavoriteActresses，應延遲到 callback 內"
-
-    def test_toggle_actress_mode_animgen_guard(self):
-        """toggleActressMode 函數體內 _animGeneration 出現 ≥ 2 次（外層 gen + 內層 gen2 race guard）"""
-        js = self._core_js()
-        body = self._extract_method_body(js, 'toggleActressMode')
-        count = len(re.findall(r'_animGeneration', body))
-        assert count >= 2, \
-            f"toggleActressMode 函數體 _animGeneration 出現次數應 ≥ 2 (外 gen + 內 gen2)，實際 {count}"
-
-    def test_old_caller_backward_compat(self):
-        """switchMode 內 playModeCrossfade 呼叫不含 onOldFadeComplete（保持影片模式內切換行為不變）。
-        searchActressFilms 自 T7 起為 async 並使用 onOldFadeComplete 觸發 ghost fly fade-out，
-        故僅驗證 switchMode 路徑不退化。"""
-        js = self._core_js()
-        search_body = self._extract_method_body(js, 'searchActressFilms')
-        switch_body = self._extract_method_body(js, 'switchMode')
-        # 兩處都應呼叫 playModeCrossfade
-        assert 'playModeCrossfade' in search_body, \
-            "searchActressFilms 應仍呼叫 playModeCrossfade"
-        assert 'playModeCrossfade' in switch_body, \
-            "switchMode 應仍呼叫 playModeCrossfade"
-        # switchMode 不該帶 onOldFadeComplete（保持原 2/3-arg 行為）
-        assert 'onOldFadeComplete' not in switch_body, \
-            "switchMode 內 playModeCrossfade 呼叫不應帶 onOldFadeComplete（保持影片模式內切換行為不變）"
-
-    def test_toggle_actress_mode_handles_animations_unavailable(self):
-        """Codex P1: animations.js 不可用時 toggleActressMode 必須有 fallback path（不能讓 callback 永不觸發）"""
-        js = self._core_js()
-        body = self._extract_method_body(js, 'toggleActressMode')
-        # callback body 應抽成 named function（給 onOldFadeComplete 用、也給 fallback path 用）
-        assert re.search(
-            r'(?:function\s+\w*FadeIn\w*|var\s+\w*FadeIn\w*\s*=\s*function|\w*FadeIn\w*\s*=\s*function)',
-            body,
-        ), "toggleActressMode 應將 callback body 抽成 named function（如 flipAndFadeIn）以便 fallback 重用"
-        # 必須顯式檢查 playModeCrossfade 是否存在（不能單靠 optional chaining 短路）
-        assert re.search(
-            r'(?:typeof\s+\w+\s*===\s*[\'"]function[\'"]|window\.ShowcaseAnimations\s*&&\s*window\.ShowcaseAnimations\.playModeCrossfade)',
-            body,
-        ), "toggleActressMode 應顯式檢查 playModeCrossfade 是否可用（不能單靠 optional chaining）"
-        # 抽出來的 named function 應在函數體內被引用 ≥ 2 次（一次給 onOldFadeComplete、一次 fallback 直接呼叫）
-        # 找出第一個 *FadeIn* 識別字
-        m = re.search(r'\b(\w*[Ff]adeIn\w*)\b', body)
-        assert m is not None, "toggleActressMode 找不到 FadeIn 命名函數"
-        fname = m.group(1)
-        count = len(re.findall(r'\b' + re.escape(fname) + r'\b', body))
-        assert count >= 3, (
-            f"toggleActressMode 內 {fname} 應出現 ≥ 3 次"
-            f"（宣告 1 + onOldFadeComplete 引用 1 + fallback 同步呼叫 1），實際 {count}"
-        )
-
-    def test_toggle_actress_mode_reduced_motion_guard_on_fade_in(self):
-        """Codex P2: toggleActressMode 內 newEl fade-in 必須有 reduced-motion 防護。
-
-        49a-T4 起：原本的 inline `gsap.fromTo` 已重構為呼叫
-        `window.ShowcaseAnimations.playContainerFadeIn`，該 helper 內部的
-        `shouldSkip()` 已涵蓋 reduced-motion。本 test 接受兩種寫法擇一：
-        (a) inline guard（舊架構）— 函數體內含 `prefersReducedMotion` 檢查
-        (b) helper 委派（新架構）— 函數體內呼叫 `playContainerFadeIn`
-        """
-        js = self._core_js()
-        body = self._extract_method_body(js, 'toggleActressMode')
-        has_inline_guard = 'prefersReducedMotion' in body
-        has_helper_delegation = 'playContainerFadeIn' in body
-        assert has_inline_guard or has_helper_delegation, (
-            "toggleActressMode newEl fade-in 應走 inline prefersReducedMotion guard，"
-            "或委派 ShowcaseAnimations.playContainerFadeIn helper（後者 shouldSkip 已涵蓋）"
-        )
 
 
-class TestAliasLiveQueryGuard:
-    """49a-T3: Actress Lightbox 別名即時查 guard
-
-    驗證：
-    - _fetchLiveAliases 方法存在
-    - 200 分支以 Object.assign 覆蓋 aliases（CD-4 + §8.4 reactivity）
-    - 404 / error / timeout 保留 snapshot（catch + 不覆蓋 fallback）
-    """
-
-    def _js(self):
-        # _fetchLiveAliases / openActressLightbox / prevActressLightbox / nextActressLightbox → state-actress.js
-        # openHeroCardLightbox → state-lightbox.js
-        return (
-            SHOWCASE_ACTRESS_JS.read_text(encoding="utf-8") + "\n" +
-            SHOWCASE_LIGHTBOX_JS.read_text(encoding="utf-8")
-        )
-
-    def _extract_method_body(self, js, method_name):
-        """抓取 Alpine state method 函式主體，大括號平衡（容忍 async 前綴）。"""
-        pattern = re.compile(
-            r'(?:^|\n)\s*(?:async\s+)?' + re.escape(method_name) + r'\s*\([^)]*\)\s*\{',
-            re.DOTALL,
-        )
-        m = pattern.search(js)
-        assert m is not None, f"showcase/core.js 找不到 {method_name} 方法"
-        start = m.end()
-        depth = 1
-        i = start
-        while i < len(js) and depth > 0:
-            c = js[i]
-            if c == '{':
-                depth += 1
-            elif c == '}':
-                depth -= 1
-            i += 1
-        return js[start:i - 1]
-
-    def test_fetch_live_aliases_method_exists(self):
-        """core.js 含 async _fetchLiveAliases 方法定義"""
-        js = self._js()
-        assert re.search(r'async\s+_fetchLiveAliases\s*\([^)]*\)\s*\{', js), \
-            "showcase/core.js 缺少 async _fetchLiveAliases(...) 方法定義"
-        # 必須呼叫 /api/actress-aliases/ 端點
-        body = self._extract_method_body(js, '_fetchLiveAliases')
-        assert "/api/actress-aliases/" in body, \
-            "_fetchLiveAliases 函數體缺少 /api/actress-aliases/ 端點呼叫"
-
-    def test_200_branch_uses_object_assign(self):
-        """200 分支用 Object.assign 覆蓋 currentLightboxActress.aliases（§8.4 Alpine reactivity）"""
-        js = self._js()
-        body = self._extract_method_body(js, '_fetchLiveAliases')
-        # 必須有 200 status 分支
-        assert re.search(r'(?:resp|response)\.status\s*===\s*200', body), \
-            "_fetchLiveAliases 缺少 resp.status === 200 分支"
-        # 必須用 Object.assign 建立新物件以觸發 Alpine deep watch（§8.4）
-        assert re.search(r'Object\.assign\s*\(', body), \
-            "_fetchLiveAliases 200 分支應用 Object.assign 建立新物件以觸發 Alpine reactivity（§8.4）"
-        # 覆蓋的目標必須是 aliases
-        assert re.search(r'aliases\s*:', body), \
-            "_fetchLiveAliases Object.assign 應指定 aliases 屬性"
-
-    def test_fallback_preserves_snapshot_on_error(self):
-        """error / timeout / 404 分支保留 snapshot（catch 區塊不覆蓋 aliases）"""
-        js = self._js()
-        body = self._extract_method_body(js, '_fetchLiveAliases')
-        # 必須有 try / catch 區塊（fallback contract）
-        assert re.search(r'\btry\s*\{', body), \
-            "_fetchLiveAliases 缺少 try 區塊（error fallback contract）"
-        assert re.search(r'\bcatch\s*\(', body), \
-            "_fetchLiveAliases 缺少 catch 區塊（error fallback contract）"
-        # 200 分支應用 if 包裹（亦即 404/其他狀態落入 implicit fallback：不執行覆蓋）
-        # 實作必須讓「非 200」 + 「catch」 不執行 Object.assign
-        # 用結構驗證：Object.assign 必須出現在 if (resp.status === 200) { ... } 區塊內
-        pattern = re.compile(
-            r'if\s*\(\s*(?:resp|response)\.status\s*===\s*200\s*\)\s*\{[^}]*?Object\.assign',
-            re.DOTALL,
-        )
-        assert pattern.search(body), \
-            "_fetchLiveAliases Object.assign 應位於 if (resp.status === 200) {...} 區塊內，避免非 200 分支誤覆蓋 snapshot"
-
-    def test_callsites_in_open_actress_and_hero(self):
-        """openActressLightbox（兩分支）+ openHeroCardLightbox 皆 fire-and-forget 呼叫 _fetchLiveAliases"""
-        js = self._js()
-        actress_body = self._extract_method_body(js, 'openActressLightbox')
-        # 至少 2 處（首次進入 + 切換女優）
-        actress_calls = re.findall(r'_fetchLiveAliases\s*\(', actress_body)
-        assert len(actress_calls) >= 2, \
-            f"openActressLightbox 應至少 2 處呼叫 _fetchLiveAliases（首次進入 + 切換女優），目前 {len(actress_calls)} 處"
-
-        hero_body = self._extract_method_body(js, 'openHeroCardLightbox')
-        assert re.search(r'_fetchLiveAliases\s*\(', hero_body), \
-            "openHeroCardLightbox 缺少 _fetchLiveAliases 呼叫"
-
-    def test_prev_next_actress_lightbox_refetch_aliases(self):
-        """Codex P2: prev/nextActressLightbox 在切換 index 後也須呼叫 _fetchLiveAliases，
-        否則方向鍵切換時 SSOT 心智模型破功（看到 stale snapshot）。"""
-        js = self._js()
-        for method in ('prevActressLightbox', 'nextActressLightbox'):
-            body = self._extract_method_body(js, method)
-            assert re.search(r'_fetchLiveAliases\s*\(', body), (
-                f"{method} 缺少 _fetchLiveAliases 呼叫（Codex P2 fix）— "
-                "方向鍵切換時不重抓 alias，違反 T3 SSOT 設計"
-            )
 
 
 GHOST_FLY_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "shared" / "ghost-fly.js"
@@ -3299,147 +2446,6 @@ class TestDetailSwipeGuard:
 SHOWCASE_CSS_T4CD = Path(__file__).parent.parent.parent / "web" / "static" / "css" / "pages" / "showcase.css"
 
 
-class TestPickerIntegrationGuard:
-    """49b-T4cd: 守衛 Actress Photo Picker 在 Showcase Lightbox 的 UI + Alpine + SSE 整合（method folded）"""
-
-    def _html(self):
-        return SHOWCASE_HTML.read_text(encoding="utf-8")
-
-    def _core_js(self):
-        return SHOWCASE_LIGHTBOX_JS.read_text(encoding="utf-8")
-
-    def _css(self):
-        return SHOWCASE_CSS_T4CD.read_text(encoding="utf-8")
-
-    def test_picker_html_contains(self):
-        """showcase.html 含 picker button、overlay 結構"""
-        html = self._html()
-        for expected in [
-            "bi-arrow-clockwise",
-            "showcase.actress.change_photo",
-            "currentLightboxActress?.is_favorite",
-            "actress-picker-overlay",
-            "picker-candidates-grid",
-            "picker-source-badge",
-            "picker-loading",
-            "picker-empty",
-        ]:
-            assert expected in html, f"showcase.html missing: {expected!r}"
-        # T1: actress-picker-area must be renamed
-        assert "actress-picker-area" not in html, \
-            "showcase.html should not contain: 'actress-picker-area'"
-
-    def test_picker_js_contains(self):
-        """core.js 含 picker state、methods、params、SSE handler 等必要字串"""
-        js = self._core_js()
-        for expected in [
-            # state
-            "_pickerOpen: false",
-            "_pickerRunId: 0",
-            "_candidates: []",
-            "_pickerSelected: false",
-            # methods
-            "openActressPicker(",
-            "_startPickerSSE(",
-            "_closePicker(",
-            "_resetPicker(",
-            "_fadeMetadataPanel(",
-            "_cancelPicker",
-            # params
-            "_PICKER_PARAMS",
-            "arcOvershoot: 1.3",
-            # burst picker animations
-            "playPickerFlipReplace",
-            "playPickerExitAll",
-            "typeof window.BurstPicker",
-            "playPickerReverseAll",
-            # SSE defer-burst
-            "_burstAllPickerCandidates",
-            # i18n
-            "showcase.actress.picker.replaced",
-            "showcase.actress.picker.error",
-            "showToast(",
-            # reduced motion
-            "prefers-reduced-motion",
-            "matchMedia",
-            # lightbox teardown
-            "_pickerOpen",
-            "_closePicker",
-            # stale name capture
-            "capturedName",
-            "currentLightboxActress",
-        ]:
-            assert expected in js, f"core.js missing: {expected!r}"
-        # arcDuration
-        assert ("arcDuration:  0.75" in js or "arcDuration: 0.75" in js), \
-            "core.js missing: 'arcDuration: 0.75' in _PICKER_PARAMS"
-        # _burstAllPickerCandidates ≥ 4 occurrences
-        assert js.count("_burstAllPickerCandidates") >= 4, \
-            "_burstAllPickerCandidates must appear ≥4 times (def + done/timeout/error)"
-
-    def test_picker_css_rules_present(self):
-        """showcase.css 含 .picker-candidate-card opacity:0 + overlay fixed + spin keyframes"""
-        css = self._css()
-        assert ".picker-candidate-card" in css, \
-            "showcase.css missing: '.picker-candidate-card'"
-        card_block = re.search(
-            r"(?:^|\n)\.picker-candidate-card\s*\{[^}]*\}", css, re.DOTALL
-        )
-        assert card_block, "showcase.css: cannot find .picker-candidate-card style block"
-        assert "opacity: 0" in card_block.group(0), \
-            ".picker-candidate-card missing: 'opacity: 0'"
-        area_block = re.search(
-            r"\.actress-picker-overlay\s*\{[^}]*\}", css, re.DOTALL
-        )
-        assert area_block, "showcase.css: cannot find .actress-picker-overlay style block"
-        overlay_css = area_block.group(0)
-        for expected in ["position: fixed", "bottom:", "width:"]:
-            assert expected in overlay_css, \
-                f".actress-picker-overlay missing: {expected!r}"
-        assert "@keyframes spin" in css, \
-            "showcase.css missing: '@keyframes spin'"
-
-    def test_picker_overlay_is_showcase_lightbox_direct_child(self):
-        """49c-T1: actress-picker-overlay 必須為 .showcase-lightbox 的直接 child"""
-        import html.parser as _html_parser
-
-        html_text = self._html()
-        assert "actress-picker-overlay" in html_text, \
-            "showcase.html missing: 'actress-picker-overlay'"
-
-        class _DivStackParser(_html_parser.HTMLParser):
-            def __init__(self):
-                super().__init__()
-                self.div_stack = []
-                self.overlay_ancestors = None
-                self.found_overlay_in_lightbox_content = False
-
-            def handle_starttag(self, tag, attrs):
-                if tag != "div":
-                    return
-                attr_dict = dict(attrs)
-                classes = set(attr_dict.get("class", "").split())
-                if "actress-picker-overlay" in classes:
-                    if self.overlay_ancestors is None:
-                        self.overlay_ancestors = [s.copy() for s in self.div_stack]
-                    if any("lightbox-content" in s for s in self.div_stack):
-                        self.found_overlay_in_lightbox_content = True
-                self.div_stack.append(classes)
-
-            def handle_endtag(self, tag):
-                if tag != "div":
-                    return
-                if self.div_stack:
-                    self.div_stack.pop()
-
-        parser = _DivStackParser()
-        parser.feed(html_text)
-        assert parser.overlay_ancestors is not None, \
-            "actress-picker-overlay not found in markup"
-        assert not parser.found_overlay_in_lightbox_content, \
-            "actress-picker-overlay should not be inside lightbox-content"
-        assert "showcase-lightbox" in parser.overlay_ancestors[-1], \
-            "actress-picker-overlay direct parent should have showcase-lightbox class"
 
 # Removed in T55b — superseded by stylelint:
 #   TestSettingsCssHardcoded, TestHelpCssHardcoded, TestDesignSystemCssHardcoded
@@ -3713,69 +2719,6 @@ class TestJellyfinFrontend:
             "settings.html 或 locales/zh_TW.json Jellyfin 圖片模式描述缺少 extrafanart/ 說明（T5b）"
 
 
-class TestOpenLocalGuard:
-    """確認 openLocal() 綁定和 open_folder() API 的結構完整性（T5a / T5b）"""
-
-    def test_open_local_in_search(self):
-        """search.html 包含 openLocal( 綁定（Detail badge + Grid overlay 兩處）"""
-        html_file = PROJECT_ROOT / "web" / "templates" / "search.html"
-        content = html_file.read_text(encoding='utf-8')
-        assert 'openLocal(' in content, \
-            "search.html 缺少 openLocal( 綁定（T5b：Detail badge + Grid overlay）"
-
-    def test_open_local_in_showcase(self):
-        """showcase.html 包含 openLocal( 綁定（Grid overlay + Lightbox 兩處）"""
-        html_file = PROJECT_ROOT / "web" / "templates" / "showcase.html"
-        content = html_file.read_text(encoding='utf-8')
-        assert 'openLocal(' in content, \
-            "showcase.html 缺少 openLocal( 綁定（T5b：Grid overlay + Lightbox）"
-
-    def test_open_local_method_exists(self):
-        """result-card.js 和 showcase/state-videos.js 均包含 openLocal(path) method 定義"""
-        result_card = PROJECT_ROOT / "web" / "static" / "js" / "pages" / "search" / "state" / "result-card.js"
-        showcase_videos = PROJECT_ROOT / "web" / "static" / "js" / "pages" / "showcase" / "state-videos.js"
-
-        rc_content = result_card.read_text(encoding='utf-8')
-        assert 'openLocal(path)' in rc_content, \
-            "result-card.js 缺少 openLocal(path) method 定義（T5b）"
-
-        sc_content = showcase_videos.read_text(encoding='utf-8')
-        assert 'openLocal(path)' in sc_content, \
-            "showcase/state-videos.js 缺少 openLocal(path) method 定義（T5b）"
-
-    def test_open_folder_pywebview_api(self):
-        """windows/pywebview_api.py 包含 def open_folder（T5a）"""
-        api_file = PROJECT_ROOT / "windows" / "pywebview_api.py"
-        content = api_file.read_text(encoding='utf-8')
-        assert 'def open_folder' in content, \
-            "pywebview_api.py 缺少 def open_folder（T5a）"
-
-    def test_no_stale_copy_local_path(self):
-        """search.html 不包含 copyLocalPath( 呼叫（確認舊 call 已清除）"""
-        html_file = PROJECT_ROOT / "web" / "templates" / "search.html"
-        content = html_file.read_text(encoding='utf-8')
-        assert 'copyLocalPath(' not in content, \
-            "search.html 仍包含 copyLocalPath( — T5b 應已將其改為 openLocal()"
-
-    def test_open_local_checks_return_value(self):
-        """openLocal() 的 .then() 必須檢查 open_folder 回傳值（不能無條件當成功）"""
-        for js_file in [
-            PROJECT_ROOT / "web" / "static" / "js" / "pages" / "search" / "state" / "result-card.js",
-            PROJECT_ROOT / "web" / "static" / "js" / "pages" / "showcase" / "state-videos.js",
-        ]:
-            content = js_file.read_text(encoding='utf-8')
-            assert '.then(async (opened)' in content, \
-                f"{js_file.name} openLocal() 的 .then() 缺少 opened 參數檢查"
-
-    def test_open_local_cross_platform_path(self):
-        """openLocal() 必須偵測 Windows drive letter 而非一律轉反斜線"""
-        for js_file in [
-            PROJECT_ROOT / "web" / "static" / "js" / "pages" / "search" / "state" / "result-card.js",
-            PROJECT_ROOT / "web" / "static" / "js" / "pages" / "showcase" / "state-videos.js",
-        ]:
-            content = js_file.read_text(encoding='utf-8')
-            assert 'displayPath' in content, \
-                f"{js_file.name} openLocal() 缺少跨平台路徑格式偵測（displayPath）"
 
 
 class TestPathContract:
@@ -4011,14 +2954,6 @@ class TestHelpPage:
             "help.js curl 模板未用 derived base（應為 `${base}/api/capabilities` — 證 data-attr 是實際來源）"
 
 
-class TestScannerClearCache:
-    """清除快取守衛 — scanner 頁面必要元素"""
-
-    def test_scanner_clear_cache_js_contains(self):
-        """scanner/state-scan.js 含 clearCache() + DELETE /api/gallery/cache"""
-        js = (PROJECT_ROOT / 'web/static/js/pages/scanner/state-scan.js').read_text(encoding='utf-8')
-        for expected in ['clearCache()', '/api/gallery/cache', 'DELETE']:
-            assert expected in js, f"scanner/state-scan.js missing: {expected!r}"
 
 
 class TestPageLifecycleGuard:
@@ -6220,130 +5155,6 @@ class TestRescrapeModalGuard:
             assert key in rescrape, f"zh_TW.json 缺少 showcase.rescrape.{key}"
 
 
-class TestRescrapeStateGuard:
-    """62a-3: 守衛 state-rescrape.js mixin contract — 確保 partial（_rescrape_modal.html）引用的
-    state/method 全揭露、commit 契約（enrich-single + refresh_full + overwrite_existing）正確、
-    transient 不碰 currentLightboxVideo（CD-62-2），且 main.js mergeState 鏈整合。
-    match TestSimilarStageGuard L1164 pattern。
-    """
-
-    SHARED_DIR = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "shared"
-    SHOWCASE_DIR = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "showcase"
-    STATE_RESCRAPE_JS = SHARED_DIR / "state-rescrape.js"
-    MAIN_JS = SHOWCASE_DIR / "main.js"
-
-    def _src(self):
-        return self.STATE_RESCRAPE_JS.read_text(encoding="utf-8")
-
-    def _main(self):
-        return self.MAIN_JS.read_text(encoding="utf-8")
-
-    def test_state_rescrape_file_exists(self):
-        """state-rescrape.js 必須存在於 shared/（供 showcase + search 共用）。"""
-        assert self.STATE_RESCRAPE_JS.exists(), \
-            f"state-rescrape.js missing at {self.STATE_RESCRAPE_JS!s}"
-
-    def test_exports_rescrape_state_factory(self):
-        """必須 export function rescrapeState（factory，rescrapeState.call(this) 接入 mergeState）。"""
-        src = self._src()
-        assert re.search(r"export\s+function\s+rescrapeState\s*\(", src), \
-            "state-rescrape.js missing: export function rescrapeState()"
-
-    def test_defines_all_state_keys(self):
-        """9 個 partial 引用的 state key 必須平鋪定義。"""
-        src = self._src()
-        for key in (
-            "rescrapeOpen", "rescrapeStep", "rescrapeEntryPoint", "rescrapeNumber",
-            "rescrapeOriginalFilename", "rescrapeSources", "rescrapeLoadingSource",
-            "rescrapePreview", "rescrapeNotFound",
-        ):
-            assert key in src, f"state-rescrape.js missing state key: {key}"
-
-    def test_defines_all_methods(self):
-        """6 個 partial 引用的 method + openRescrape（62b-1 呼叫）必須揭露。"""
-        src = self._src()
-        for method in (
-            "openRescrape", "rescrapeWithSource", "rescrapeConfirm",
-            "rescrapeBackToPick", "closeRescrape", "rescrapeBuiltinSources",
-            "rescrapeMetatubeSources",
-        ):
-            assert method in src, f"state-rescrape.js missing method: {method}"
-
-    def test_commit_contract(self):
-        """commit 契約：POST /api/enrich-single + mode refresh_full + overwrite_existing。"""
-        src = self._src()
-        assert "'/api/enrich-single'" in src, "missing POST /api/enrich-single"
-        assert "refresh_full" in src, "missing mode: refresh_full（CD-62-0/4）"
-        assert "overwrite_existing" in src, "missing overwrite_existing（CD-62-4）"
-        assert "true" in src, "overwrite_existing 必須為 true"
-
-    def test_preview_contract(self):
-        """preview 契約：POST /api/rescrape/preview。"""
-        src = self._src()
-        assert "'/api/rescrape/preview'" in src, "missing POST /api/rescrape/preview"
-
-    def test_no_current_lightbox_video(self):
-        """CD-62-2 transient 守衛：mixin 絕不讀寫 currentLightboxVideo（用私有 _rescrapeVideo）。"""
-        src = self._src()
-        assert "currentLightboxVideo" not in src, \
-            "state-rescrape.js 違反 CD-62-2：不得引用 currentLightboxVideo，應用私有 _rescrapeVideo"
-
-    def test_rescraping_guard_present(self):
-        """連點防護：_rescraping guard 必須存在（鏡像 _enriching）。"""
-        src = self._src()
-        assert "_rescraping" in src, "missing _rescraping 連點 guard"
-
-    def test_no_proxy_image_construction(self):
-        """CD-62-14 #8：/api/proxy-image URL 由 partial 內聯，mixin 不得重複建構。"""
-        src = self._src()
-        assert "/api/proxy-image" not in src, \
-            "state-rescrape.js 不得建構 /api/proxy-image URL（partial 內聯，避免重複）"
-
-    def test_main_js_imports_and_merges_rescrape_state(self):
-        """main.js 必須 import rescrapeState 並插入 mergeState 鏈。"""
-        src = self._main()
-        assert "from '@/shared/state-rescrape.js'" in src, \
-            "main.js missing: import { rescrapeState } from '@/shared/state-rescrape.js'"
-        assert "rescrapeState.call(this)" in src, \
-            "main.js mergeState chain missing: rescrapeState.call(this)"
-
-    def test_open_rescrape_reads_video_number(self):
-        """62b-2 #1：openRescrape 內 rescrapeNumber 預填來源必須 = video.number（前端 prefill 持久化連結）。
-
-        鎖住「commit 修正 number → refreshVideoData 突變 video.number → 再開彈窗預填新值」鏈的前端端點。
-        若 refactor 把預填改成讀別處（如固定 '' 或 video.code），此守衛 RED。
-        """
-        src = self._src()
-        # 寬鬆匹配：rescrapeNumber = (... video.number ...) ；允許 =、&&、() 周圍空白變動
-        assert re.search(r"rescrapeNumber\s*=.*video\s*&&\s*video\.number", src), \
-            "openRescrape 必須將 rescrapeNumber 預填自 video.number（前端 prefill 連結，62b-2 #6）"
-
-    def test_close_rescrape_clears_longpress_flag(self):
-        """74c-T3（翻轉）：longPressReset?.() 呼叫已隨長壓基礎設施退役移除；state-rescrape.js 不得再含此呼叫。"""
-        src = self._src()
-        assert "longPressReset" not in src, \
-            "74c-T3 違規：closeRescrape 仍含 longPressReset（長壓基礎設施已退役，此呼叫應移除）"
-
-    def test_rescrape_metatube_sources_has_routable_gate(self):
-        """Codex PR#47 round-2 P2-B：rescrapeMetatubeSources() 必須同時 filter
-        type === 'metatube' AND routable === true（element-bound regex，防空字串假測試）。
-
-        metatube sources 目前後端無路由（validate_source_id 只認 auto + builtin SOURCE_ORDER）；
-        B3 才接 metatube route/validator，屆時後端揭露 routable=true 後 pill 才長出。
-        直到 B3 前，缺 routable gate 的 metatube pill 點下去 → not-found；本守衛確保回歸會 RED。
-        """
-        src = self._src()
-        # element-bound regex：匹配 rescrapeMetatubeSources 函式體，確認同時含兩個 filter 條件
-        m = re.search(
-            r"rescrapeMetatubeSources\s*\(\s*\)\s*\{[^}]*\.filter\s*\([^)]*"
-            r"s\.type\s*===\s*['\"]metatube['\"][^)]*&&[^)]*s\.routable\s*===\s*true[^)]*\)",
-            src,
-            re.DOTALL,
-        )
-        assert m, (
-            "rescrapeMetatubeSources() 必須 filter s.type === 'metatube' && s.routable === true "
-            "（缺 routable gate：metatube pill 在後端無路由前點下去 → not-found，Codex PR#47 round-2 P2-B）"
-        )
 
 
 class TestRescrapeVersionStateGuard:
@@ -8256,128 +7067,6 @@ class TestCoverLoadingUx67Guard:
             "_setLightboxIndex 未委託 _refreshLbFullBlurUp（71c-P2 抽 helper 後應呼叫 helper 不可 inline）"
 
 
-class TestPartsBinStagedAffordanceGuard:
-    """Parts Bin 可達/不可達膠囊視覺語義對調守衛（TASK-partsbin-staged-affordance）。
-
-    Template（settings.html）↔ CSS（source-pill.css）↔ Design-system（D.13）跨檔 contract；
-    eslint 不解析 Jinja template，stylelint 無法表達選擇器歸屬語義，故走 pytest
-    （沿用 TestPicker64aThreeStateGuard / TestCoverLoadingUx67Guard 先例）。
-
-    強度：先 regex 擷取目標區塊再斷言，非整檔裸字串存在性（gotchas G5）。
-    6 條契約各對應 1 個 test method，獨立失敗便於定位。
-    """
-
-    SETTINGS_HTML    = Path(__file__).parent.parent.parent / "web" / "templates" / "settings.html"
-    SOURCE_PILL_CSS  = Path(__file__).parent.parent.parent / "web" / "static" / "css" / "components" / "source-pill.css"
-    DS_SETTINGS_HTML = Path(__file__).parent.parent.parent / "web" / "templates" / "design_system" / "settings-components.html"
-
-    def _settings(self):
-        return self.SETTINGS_HTML.read_text(encoding="utf-8")
-
-    def _css(self):
-        return self.SOURCE_PILL_CSS.read_text(encoding="utf-8")
-
-    def _ds(self):
-        return self.DS_SETTINGS_HTML.read_text(encoding="utf-8")
-
-    def _partsbin_pill_block(self):
-        """抽出 settings.html Parts Bin pill loop 的 template x-for 區塊。"""
-        html = self._settings()
-        m = re.search(
-            r'<template x-for="src in partsBinSources"[^>]*>.*?</template>',
-            html, re.DOTALL,
-        )
-        assert m, "settings.html: 找不到 Parts Bin pill x-for 區塊（partsBinSources）"
-        return m.group(0)
-
-    # ── 契約 1：settings.html Parts Bin pill 含 bi-plus-circle（plus-icn）──────────────
-
-    def test_settings_partsbin_pill_has_plus_icn(self):
-        """Contract 1：settings.html Parts Bin pill loop 內存在 bi-plus-circle plus-icn（可加入 affordance）。"""
-        block = self._partsbin_pill_block()
-        assert "bi-plus-circle" in block, (
-            "TASK-partsbin 違規 C1：settings.html Parts Bin pill 缺 bi-plus-circle（plus-icn）。"
-            "可加入 affordance 需常駐 DOM，靠 CSS 依 data-available 控顯隱。"
-        )
-        assert "plus-icn" in block, (
-            "TASK-partsbin 違規 C1：settings.html Parts Bin pill 缺 plus-icn class。"
-        )
-
-    # ── 契約 2：settings.html 不再含 settings-mt-probe-hint / mt_probe_hint_title ──────
-
-    def test_settings_no_probe_hint_details(self):
-        """Contract 2：settings.html 不含 settings-mt-probe-hint（3-cause <details> 已硬刪）。"""
-        html = self._settings()
-        assert "settings-mt-probe-hint" not in html, (
-            "TASK-partsbin 違規 C2：settings.html 仍含 settings-mt-probe-hint（三因摺疊區塊應已整段硬刪）。"
-        )
-        assert "mt_probe_hint_title" not in html, (
-            "TASK-partsbin 違規 C2：settings.html 仍含 mt_probe_hint_title（三因摺疊標題應已整段硬刪）。"
-        )
-
-    # ── 契約 3：settings.html Parts Bin slash-icn 仍在（不可達態靠它）────────────────
-
-    def test_settings_partsbin_pill_slash_icn_retained(self):
-        """Contract 3：settings.html Parts Bin pill loop 內 slash-icn 仍在（不可達態靠 CSS 顯示）。"""
-        block = self._partsbin_pill_block()
-        assert "slash-icn" in block, (
-            "TASK-partsbin 違規 C3：settings.html Parts Bin pill 的 slash-icn 被誤刪（不可達態靠它，CSS 控顯隱）。"
-        )
-        assert "bi-slash-circle" in block, (
-            "TASK-partsbin 違規 C3：settings.html Parts Bin pill 缺 bi-slash-circle icon。"
-        )
-
-    # ── 契約 4：source-pill.css 含可達態 .pill-name text-decoration:none ───────────────
-
-    def test_css_partsbin_available_true_removes_line_through(self):
-        """Contract 4：source-pill.css 含 .is-partsbin[data-available='true'] .pill-name text-decoration:none（選擇器歸屬）。"""
-        css = self._css()
-        m = re.search(
-            r'\.source-pill\.is-partsbin\[data-available="true"\]\s+\.pill-name\s*\{([^}]+)\}',
-            css, re.DOTALL,
-        )
-        assert m, (
-            "TASK-partsbin 違規 C4：source-pill.css 缺 "
-            ".source-pill.is-partsbin[data-available=\"true\"] .pill-name 規則。"
-            "（可達正面態需 (0,4,0) specificity 勝全域 (0,3,0) line-through）"
-        )
-        assert "text-decoration: none" in m.group(1), (
-            "TASK-partsbin 違規 C4：可達態 .pill-name 規則缺 text-decoration:none。"
-        )
-
-    # ── 契約 5：source-pill.css .is-partsbin cursor pointer ───────────────────────────
-
-    def test_css_partsbin_cursor_pointer(self):
-        """Contract 5：source-pill.css .source-pill.is-partsbin cursor 為 pointer（click-to-promote 語義）。"""
-        css = self._css()
-        m = re.search(
-            r'\.source-pill\.is-partsbin\s*\{([^}]+)\}',
-            css, re.DOTALL,
-        )
-        assert m, (
-            "TASK-partsbin 違規 C5：source-pill.css 缺 .source-pill.is-partsbin cursor 規則。"
-        )
-        assert "cursor: pointer" in m.group(1), (
-            "TASK-partsbin 違規 C5：.source-pill.is-partsbin 的 cursor 非 pointer。"
-        )
-
-    # ── 契約 6：design-system D.13 不含 rec-star/data-rec，含 available true/false 兩 demo ──
-
-    def test_ds_d13_no_rec_star_and_has_both_available_states(self):
-        """Contract 6：design-system D.13 不含 rec-star/data-rec（dead 清除），含 available=true 與 false 兩 demo。"""
-        ds = self._ds()
-        assert "rec-star" not in ds, (
-            "TASK-partsbin 違規 C6：design-system settings-components.html 仍含 rec-star（dead CSS class，應清除）。"
-        )
-        assert "data-rec" not in ds, (
-            "TASK-partsbin 違規 C6：design-system settings-components.html 仍含 data-rec 屬性（dead，應清除）。"
-        )
-        assert 'data-available="true"' in ds, (
-            "TASK-partsbin 違規 C6：design-system D.13 缺 is-partsbin data-available=\"true\" demo（staged 可加入態）。"
-        )
-        assert 'data-available="false"' in ds, (
-            "TASK-partsbin 違規 C6：design-system D.13 缺 is-partsbin data-available=\"false\" demo（不可達態）。"
-        )
 
 
 # ── TASK-70-T5: JavLibrary Picker BETA 視覺 + 不可用 gate 靜態守衛 ──
@@ -8389,73 +7078,6 @@ _MODAL_HTML_70 = Path(__file__).parent.parent.parent / "web" / "templates" / "_r
 _LOCALES_ROOT_70 = Path(__file__).parent.parent.parent / "locales"
 
 
-class TestJavlibraryPickerT5Guard:
-    """70-T5: javlibrary picker BETA 視覺 + 不可用 gate 靜態守衛。
-
-    守衛靜態字串契約（HTML/JS/Python）：
-      (1) bootstrap template 含 cf_transport_available 注入
-      (2) app.py get_common_context 注入 cf_transport_available
-      (3) state-rescrape.js 定義 isJlUnavailable
-      (4) _rescrape_modal.html builtin pill 含 source-pill-badge（BETA）綁於 manual_only && is_beta
-      (5) _rescrape_modal.html builtin pill 含 isJlUnavailable gate（aria-disabled + jl_desktop_only toast）
-      (6) 4 locale 檔含 jl_desktop_only key
-    """
-
-    def test_bootstrap_cf_transport_available(self):
-        """(1) _advanced_search_bootstrap.html 含 cf_transport_available 注入。"""
-        html = _BOOTSTRAP_HTML.read_text(encoding="utf-8")
-        assert "cf_transport_available" in html, (
-            "70-T5 違規：_advanced_search_bootstrap.html 未注入 cf_transport_available 到 __ADVANCED_SEARCH__"
-        )
-
-    def test_app_py_get_common_context_cf_transport(self):
-        """(2) web/app.py get_common_context() 注入 cf_transport_available。"""
-        src = _APP_PY.read_text(encoding="utf-8")
-        assert "cf_transport_available" in src, (
-            "70-T5 違規：web/app.py get_common_context() 未包含 cf_transport_available key"
-        )
-        assert "get_cf_transport" in src, (
-            "70-T5 違規：web/app.py 未 import 或呼叫 get_cf_transport（cf_transport_available 值的來源）"
-        )
-
-    def test_state_rescrape_has_isJlUnavailable(self):
-        """(3) state-rescrape.js 定義 isJlUnavailable helper。"""
-        js = _STATE_RESCRAPE_JS.read_text(encoding="utf-8")
-        assert "isJlUnavailable" in js, (
-            "70-T5 違規：state-rescrape.js 未定義 isJlUnavailable helper"
-        )
-
-    def test_modal_builtin_pill_has_beta_badge(self):
-        """(4) _rescrape_modal.html builtin pill 模板含 source-pill-badge（BETA badge）。"""
-        html = _MODAL_HTML_70.read_text(encoding="utf-8")
-        assert "source-pill-badge" in html, (
-            "70-T5 違規：_rescrape_modal.html builtin pill 模板缺 source-pill-badge（BETA badge）"
-        )
-        assert "manual_only" in html and "is_beta" in html, (
-            "70-T5 違規：source-pill-badge 未綁於 manual_only && is_beta 條件（不應對所有 pill 顯示）"
-        )
-
-    def test_modal_builtin_pill_jl_unavailable_gate(self):
-        """(5) _rescrape_modal.html builtin pill 含 isJlUnavailable gate（aria-disabled + jl_desktop_only）。"""
-        html = _MODAL_HTML_70.read_text(encoding="utf-8")
-        assert "isJlUnavailable" in html, (
-            "70-T5 違規：_rescrape_modal.html builtin pill 模板未使用 isJlUnavailable gate"
-        )
-        assert "aria-disabled" in html, (
-            "70-T5 違規：_rescrape_modal.html builtin pill 缺 aria-disabled 綁定（不可用語義）"
-        )
-        assert "jl_desktop_only" in html, (
-            "70-T5 違規：_rescrape_modal.html builtin pill 缺 jl_desktop_only i18n key（toast / title）"
-        )
-
-    def test_i18n_jl_desktop_only_parity(self):
-        """(6) 4 locale 檔（zh_TW/zh_CN/en/ja）皆含 jl_desktop_only key。"""
-        for lang in ("zh_TW", "zh_CN", "en", "ja"):
-            locale_path = _LOCALES_ROOT_70 / f"{lang}.json"
-            content = locale_path.read_text(encoding="utf-8")
-            assert "jl_desktop_only" in content, (
-                f"70-T5 違規：locales/{lang}.json 缺 showcase.rescrape.jl_desktop_only key（i18n parity）"
-            )
 
 
 # ── TASK-70-T6: CF flow 前端靜態守衛 ──
@@ -8862,211 +7484,12 @@ STATE_RESCRAPE_JS = (
 )
 
 
-class TestCfPollUnavailableGuard:
-    """
-    CD-70c-3: C-class API contract guard.
-
-    The frontend _pollCfThenRetry() must consume the backend's
-    {unavailable: true} signal by:
-      1. Checking data.unavailable in the poll callback
-      2. Calling cancelCfPoll() to stop polling and emit the abandon notification
-
-    This is a cross-boundary contract: cf.py response shape × frontend consumption.
-    Per CLAUDE.md: "兩個檔案之間的 API contract → pytest (C/E 類)"
-    """
-
-    def _js(self) -> str:
-        return STATE_RESCRAPE_JS.read_text(encoding="utf-8")
-
-    def test_poll_checks_data_unavailable(self):
-        """
-        state-rescrape.js _pollCfThenRetry() must check data.unavailable.
-        This is the frontend consumer of {ready: false, unavailable: true}
-        from /api/cf/status when the transport window is dead.
-        """
-        js = self._js()
-        assert "data.unavailable" in js, (
-            "state-rescrape.js _pollCfThenRetry() missing data.unavailable check — "
-            "CD-70c-3: frontend must read the backend unavailable signal to stop polling"
-        )
-
-    def test_poll_calls_cancel_cf_poll_on_unavailable(self):
-        """
-        The data.unavailable branch itself must call cancelCfPoll() (not merely
-        have cancelCfPoll defined elsewhere in the file). Verify cancelCfPoll
-        appears within a tight window right after the data.unavailable check —
-        this pins the branch wiring, not just string presence.
-        cancelCfPoll() does clearInterval + POST /api/cf/abandon (emits notification).
-        """
-        js = self._js()
-        idx = js.find("data.unavailable")
-        assert idx != -1, (
-            "state-rescrape.js missing data.unavailable check — "
-            "CD-70c-3: frontend must read the backend unavailable signal"
-        )
-        branch = js[idx: idx + 200]  # the unavailable branch body, right after the check
-        assert "cancelCfPoll" in branch, (
-            "data.unavailable branch does not call cancelCfPoll() — "
-            "CD-70c-3: dead transport must immediately stop the poll loop "
-            "(clearInterval + abandon), not fall through to the timeout"
-        )
-
-    def test_unavailable_check_present_in_poll_interval(self):
-        """
-        Both data.unavailable and cancelCfPoll appear in the same file and are
-        co-located in the polling context (not just in separate unrelated methods).
-        Verify both appear within the _pollCfThenRetry function definition text span.
-        """
-        js = self._js()
-        # Find the function *definition* (not a call site) by searching for the
-        # pattern "methodName(" preceded by whitespace/newline (method definition form).
-        import re as _re
-        # Match the function definition: optional whitespace then _pollCfThenRetry(
-        m = _re.search(r'\b_pollCfThenRetry\s*\(', js)
-        assert m is not None, f"state-rescrape.js missing _pollCfThenRetry function"
-        # Scan from the first match: if it's a call site (short arg like 'number'),
-        # try finding the definition via "function body" marker (contains 'setInterval').
-        # Use rfind to find the last definition (definitions come after call sites).
-        all_matches = list(_re.finditer(r'\b_pollCfThenRetry\s*\(', js))
-        # Prefer the match followed by a simple parameter name (definition) over
-        # a call expression with 'this.' arguments.  The definition looks like:
-        #   _pollCfThenRetry(number) {
-        def_idx = None
-        for match in all_matches:
-            tail = js[match.start(): match.start() + 80]
-            # Definition has a plain identifier parameter, not 'this.'
-            if _re.match(r'\b_pollCfThenRetry\s*\(\s*\w+\s*\)\s*\{', tail):
-                def_idx = match.start()
-                break
-        assert def_idx is not None, (
-            "Could not find _pollCfThenRetry(param) { definition in state-rescrape.js"
-        )
-        # Extract ~1500 chars from the function start to cover the setInterval body
-        snippet = js[def_idx: def_idx + 1500]
-        assert "data.unavailable" in snippet, (
-            "data.unavailable check not found inside _pollCfThenRetry() definition — "
-            "the check must be inside the setInterval callback"
-        )
-        assert "cancelCfPoll" in snippet, (
-            "cancelCfPoll() call not found inside _pollCfThenRetry() definition — "
-            "must be called when data.unavailable is true"
-        )
 
 
 SHOWCASE_CSS = Path(__file__).parent.parent.parent / "web" / "static" / "css" / "pages" / "showcase.css"
 SHOWCASE_SIMILAR_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "showcase" / "state-similar.js"
 
 
-class TestLightboxCoverSizeGuards:
-    """71c: 守衛 lightbox 封面縮水修復（thumb 放大填滿 + blur-up 鏡像 + same-URL complete-check）
-
-    四條 element-bound 守衛：
-    G1 — .lightbox-cover img 有明確 height:（非僅 max-height），確保 400px thumb 放大到 60vh
-    G2 — .lb-full 有明確 height:（非僅 max-height），確保 overlay 與 base 尺寸鏡像
-    G3 — state-lightbox.js 含 _refreshLbFullBlurUp helper（same-URL complete-check 抽出共用），
-         且 _setLightboxIndex 與 slip-through 兩處均呼叫該 helper（防回歸繞過）
-    G4 — state-similar.js similarExitVideo 含 cover_full_url 欄位
-    """
-
-    def _css(self):
-        return SHOWCASE_CSS.read_text(encoding="utf-8")
-
-    def _lightbox_js(self):
-        return SHOWCASE_LIGHTBOX_JS.read_text(encoding="utf-8")
-
-    def _similar_js(self):
-        return SHOWCASE_SIMILAR_JS.read_text(encoding="utf-8")
-
-    def _html(self):
-        return SHOWCASE_HTML.read_text(encoding="utf-8")
-
-    def test_lightbox_cover_img_has_explicit_height(self):
-        """G1: .lightbox-cover img 含明確 height: 規則（非僅 max-height），讓 thumb 放大填滿"""
-        css = self._css()
-        import re
-        # 找到 .lightbox-cover img { ... } 區塊
-        block_match = re.search(r'\.lightbox-cover\s+img\s*\{([^}]+)\}', css, re.DOTALL)
-        assert block_match, ".lightbox-cover img 規則在 showcase.css 找不到"
-        block = block_match.group(1)
-        # 確認有明確的 height:（不只是 max-height:）
-        assert re.search(r'(?<!\w)height\s*:', block), (
-            ".lightbox-cover img 缺少明確 height: 規則（只有 max-height 不足以放大小 thumb）"
-        )
-
-    def test_lb_full_has_explicit_height(self):
-        """G2: .lb-full 含明確 height: 規則（非僅 max-height），與 base img 鏡像對齊"""
-        css = self._css()
-        import re
-        block_match = re.search(r'\.lb-full\s*\{([^}]+)\}', css, re.DOTALL)
-        assert block_match, ".lb-full 規則在 showcase.css 找不到"
-        block = block_match.group(1)
-        assert re.search(r'(?<!\w)height\s*:', block), (
-            ".lb-full 缺少明確 height: 規則（overlay 需與 base img 尺寸完全鏡像）"
-        )
-
-    def test_lightbox_js_has_sameurl_complete_check(self):
-        """G3: state-lightbox.js 抽出 _refreshLbFullBlurUp helper 含 same-URL complete-check，
-        且 _setLightboxIndex 與 state-similar.js slip-through 均呼叫該 helper（71c-P2 防回歸）"""
-        import re
-        js = self._lightbox_js()
-        similar_js = self._similar_js()
-
-        # (a) helper 函數體必須存在於 state-lightbox.js，且含 complete-check 三要素
-        helper_idx = js.find("_refreshLbFullBlurUp(")
-        assert helper_idx != -1, (
-            "state-lightbox.js 找不到 _refreshLbFullBlurUp helper"
-            "（71c-P2：blur-up reset + same-URL complete-check 應抽成共用 helper）"
-        )
-        # 截取 helper 函數體（含至 closing brace，取 600 字元已足夠）
-        helper_snippet = js[helper_idx: helper_idx + 600]
-        assert "lightboxCoverFull" in helper_snippet, (
-            "_refreshLbFullBlurUp helper 缺少 lightboxCoverFull x-ref 取用"
-            "（same-URL complete-check 需 $refs.lightboxCoverFull）"
-        )
-        # 鎖 runtime 表達式（非註解）
-        assert "fullImg.complete" in helper_snippet and "fullImg.naturalWidth" in helper_snippet, (
-            "_refreshLbFullBlurUp helper 缺少 fullImg.complete && fullImg.naturalWidth 檢查"
-            "（same-URL 時瀏覽器不重新 fire @load，需手動偵測 complete）"
-        )
-        assert "_lbFullLoaded" in helper_snippet, (
-            "_refreshLbFullBlurUp helper 缺少 _lbFullLoaded 賦值"
-        )
-
-        # (b) _setLightboxIndex 必須呼叫 _refreshLbFullBlurUp（不再 inline）
-        set_idx = js.find("_setLightboxIndex(")
-        assert set_idx != -1, "state-lightbox.js 找不到 _setLightboxIndex 函數"
-        set_snippet = js[set_idx: set_idx + 800]
-        assert "_refreshLbFullBlurUp" in set_snippet, (
-            "state-lightbox.js _setLightboxIndex 未呼叫 _refreshLbFullBlurUp"
-            "（抽 helper 後 _setLightboxIndex 應委託 helper，避免邏輯漂移）"
-        )
-
-        # (c) slip-through 路徑（state-similar.js）必須在 currentLightboxVideo = similarExitVideo 之後
-        #     呼叫 _refreshLbFullBlurUp（防止繞過 _setLightboxIndex 殘留舊 _lbFullLoaded）
-        slip_idx = similar_js.find("this.currentLightboxVideo = this.similarExitVideo")
-        assert slip_idx != -1, (
-            "state-similar.js 找不到 currentLightboxVideo = similarExitVideo 指派"
-            "（slip-through 路徑應設 currentLightboxVideo 後呼叫 _refreshLbFullBlurUp）"
-        )
-        # 截取指派後 300 字元（應含 helper 呼叫）
-        after_assign = similar_js[slip_idx: slip_idx + 300]
-        assert "_refreshLbFullBlurUp" in after_assign, (
-            "state-similar.js slip-through（currentLightboxVideo = similarExitVideo）後缺少 _refreshLbFullBlurUp 呼叫"
-            "（71c-P2：slip-through 繞過 _setLightboxIndex，舊 _lbFullLoaded true → 跳過 blur-up；"
-            "false + same-URL → @load 不 fire → opacity:0 卡死）"
-        )
-
-    def test_similar_exit_video_has_cover_full_url(self):
-        """G4（JS contract）: state-similar.js similarExitVideo 含 cover_full_url 欄位"""
-        js = self._similar_js()
-        # 找 similarExitVideo = { ... } 構建區塊
-        idx = js.find("this.similarExitVideo = {")
-        assert idx != -1, "state-similar.js 找不到 similarExitVideo = { 構建"
-        snippet = js[idx: idx + 600]
-        assert "cover_full_url" in snippet, (
-            "state-similar.js similarExitVideo 缺少 cover_full_url 欄位"
-            "（slip-through 路徑缺此欄 → .lb-full src=undefined → @load 永不 fire → opacity:0 卡死）"
-        )
 
 
 SOURCE_PILL_MACRO = (
@@ -9248,48 +7671,6 @@ STATE_RESCRAPE_JS = (
 )
 
 
-class TestSourcePillFlatCss:
-    """TASK-74b-T1: .source-pill--flat 唯讀變體 CSS contract（cross-file，element-bound）。
-
-    flat 變體保留 tint、只關互動（CD-74b-1）：cursor default + hover 無 lift/shadow + focus 無 outline。
-    cross-file 契約：macro 在 variant='flat' 輸出 source-pill--flat ↔ CSS 必有對應規則。
-    過「三問」：刪 .source-pill--flat 規則 → 紅；macro 移除 flat 分支 → 紅；改 cursor 值 → 紅。
-    """
-
-    def _css(self) -> str:
-        return SOURCE_PILL_CSS.read_text(encoding="utf-8")
-
-    def _macro(self) -> str:
-        return SOURCE_PILL_MACRO.read_text(encoding="utf-8")
-
-    def test_flat_rule_defines_cursor_default(self):
-        """`.source-pill--flat { cursor: default; }` 存在（覆寫 base cursor: grab）。"""
-        css = self._css()
-        m = re.search(r"\.source-pill--flat\s*\{([^}]*)\}", css)
-        assert m, "source-pill.css 缺 .source-pill--flat 規則（74b T1 enabler）"
-        assert "cursor: default" in m.group(1), (
-            f".source-pill--flat 必須 cursor: default（關掉 base grab）；body: {m.group(1)!r}"
-        )
-
-    def test_flat_hover_negates_lift(self):
-        """`.source-pill--flat:hover` 關掉 base hover 的 transform + box-shadow。"""
-        css = self._css()
-        m = re.search(r"\.source-pill--flat:hover\s*\{([^}]*)\}", css)
-        assert m, "source-pill.css 缺 .source-pill--flat:hover 規則"
-        body = m.group(1)
-        assert "transform: none" in body, (
-            f".source-pill--flat:hover 必須 transform: none（關掉 base lift）；body: {body!r}"
-        )
-        assert "box-shadow: none" in body, (
-            f".source-pill--flat:hover 必須 box-shadow: none（關掉 base shadow）；body: {body!r}"
-        )
-
-    def test_macro_emits_flat_class_cross_file(self):
-        """cross-file：macro variant='flat' 分支輸出 source-pill--flat（CSS 規則的唯一消費路徑）。"""
-        macro = self._macro()
-        assert "source-pill--flat" in macro, (
-            "source_pill.html 未輸出 source-pill--flat — flat CSS 將無消費者（cross-file 契約斷裂）"
-        )
 
 
 class TestRescrapePreviewSourcePill:
@@ -9554,56 +7935,6 @@ GHOST_FLY_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "
 STATE_LIGHTBOX_JS = Path(__file__).parent.parent.parent / "web" / "static" / "js" / "pages" / "showcase" / "state-lightbox.js"
 
 
-class TestUS5PosterCropGhostCrossfade:
-    """TASK-75b-T7：poster 格開燈箱的 cover→contain 溶接（Codex 視覺 bug2）。
-
-    契約：state-lightbox.js 依「≤480px ∩ 非女優模式 ∩ 非 hero」算 posterCrop 並傳給
-    playGridToLightbox；ghost-fly.js 在 posterCrop 下對齊縮圖右裁（objectPosition right center）
-    並於落地 crossfade（coverEl 淡入 + ghost 淡出 0.12s）取代硬切 cleanupGhost。
-    三問：刪 posterCrop 傳遞 → 紅；刪 objectPosition 對齊 → 紅；把 crossfade 改回硬切 → 紅。
-    """
-
-    def _grid_to_lightbox_body(self) -> str:
-        js = GHOST_FLY_JS.read_text(encoding="utf-8")
-        start = js.find("playGridToLightbox: function")
-        assert start >= 0, "ghost-fly.js 找不到 playGridToLightbox"
-        end = js.find("playLightboxToGrid: function", start)
-        assert end > start, "ghost-fly.js 找不到 playGridToLightbox 結束邊界"
-        return js[start:end]
-
-    def test_state_lightbox_threads_poster_crop(self):
-        js = STATE_LIGHTBOX_JS.read_text(encoding="utf-8")
-        # 計算條件三要素
-        assert "posterCrop" in js, "state-lightbox.js 應計算 posterCrop"
-        # T11（US-10）：門檻由 ≤480 擴到 ≤899（共用常數 POSTER_CROP_MAX_W，對齊守衛 TestPosterCropThresholdAlignment）。
-        assert "window.innerWidth <= POSTER_CROP_MAX_W" in js, "posterCrop 應 gate ≤POSTER_CROP_MAX_W"
-        assert "showFavoriteActresses" in js, "posterCrop 應排除女優模式"
-        assert "hero-card" in js, "posterCrop 應排除 hero 卡（女優入口）"
-        # 傳入 playGridToLightbox 的 options
-        assert "posterCrop: posterCrop" in js, (
-            "state-lightbox.js 應把 posterCrop 傳入 playGridToLightbox options"
-        )
-
-    def test_ghost_fly_consumes_and_aligns_crop(self):
-        body = self._grid_to_lightbox_body()
-        assert "options.posterCrop" in body, "playGridToLightbox 應消費 options.posterCrop"
-        # (A) 對齊縮圖右裁
-        assert "objectPosition = 'right center'" in body, (
-            "posterCrop 下 ghost 應對齊縮圖 objectPosition right center（消起飛 pan）"
-        )
-
-    def test_ghost_fly_landing_crossfade(self):
-        body = self._grid_to_lightbox_body()
-        # (D) 落地 crossfade：coverEl 淡入 + ghost 淡出，且綁在 posterCrop 分支
-        assert "posterCrop && coverEl" in body, (
-            "落地 crossfade 應 gate 在 posterCrop（非 poster 路徑維持硬切 cleanupGhost）"
-        )
-        assert "opacity: 1, duration: 0.12" in body, "coverEl 應 0.12s 淡入（contain 真圖浮現）"
-        assert "opacity: 0, duration: 0.12" in body, "ghost 應 0.12s 淡出（溶接 cover→contain）"
-        # 非 poster 仍走硬切 cleanupGhost
-        assert "cleanupGhost(ghost, coverEl)" in body, (
-            "非 posterCrop 路徑應保留硬切 cleanupGhost（桌面零回歸）"
-        )
 
 
 # ============================================================================
@@ -9747,109 +8078,6 @@ SHOWCASE_SIMILAR_JS = (
 )
 
 
-class TestMobileSimilarDrillFallbackGuard:
-    """BUGfix-mobile-similar-stale-cover: 守衛 mobile similar drill 三層 fallback 合約
-
-    (a) state-lightbox.js 的 _setLightboxIndex 函式體必須清除 similarExitVideo（= null），
-        確保 in-grid 切換後不殘留 standalone 旗標（Codex P2b 根治）。
-    (b) state-similar.js 的 onSimilarMobileCardClick 函式體必須含：
-        - `_videos` 查找（tier 2 fallback）
-        - 設置 `similarExitVideo`（standalone 旗標）
-        - 呼叫 `_refreshLbFullBlurUp`（blur-up reset）
-    """
-
-    def _lightbox_js(self):
-        return SHOWCASE_LIGHTBOX_JS.read_text(encoding="utf-8")
-
-    def _similar_js(self):
-        return SHOWCASE_SIMILAR_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _extract_function_body(js, func_pattern):
-        """用計數括弧深度方式擷取函式體，從 func_pattern regex 匹配處起算。"""
-        m = re.search(func_pattern, js)
-        if not m:
-            return None
-        start = m.start()
-        body_start = js.index('{', start)
-        depth = 0
-        for i, ch in enumerate(js[body_start:], body_start):
-            if ch == '{':
-                depth += 1
-            elif ch == '}':
-                depth -= 1
-                if depth == 0:
-                    return js[body_start:i + 1]
-        return None
-
-    def test_set_lightbox_index_clears_similar_exit_video(self):
-        """_setLightboxIndex 函式體必須含 similarExitVideo = null（清除 standalone 旗標）。
-        根治：in-grid 切換（tier 1）後不殘留 standalone，prev/next + fly-back 恢復正常。
-        """
-        js = self._lightbox_js()
-        body = self._extract_function_body(js, r'_setLightboxIndex\s*\(')
-        assert body is not None, \
-            "state-lightbox.js 找不到 _setLightboxIndex 函式宣告"
-        # 允許有無空格兩種寫法
-        assert re.search(r'similarExitVideo\s*=\s*null', body), (
-            "_setLightboxIndex 函式體未含 'similarExitVideo = null'。\n"
-            "in-grid 切換（tier 1）後 similarExitVideo 不清除，\n"
-            "連點 tier2/3 再 tier1 時 standalone 旗標殘留，prev/next 被錯誤禁用。\n"
-            f"當前函式體：\n{body[:400]}"
-        )
-
-    def test_close_similar_mode_fallback_has_videos_tier(self):
-        """closeSimilarMode 退場 fallback 必須與 onSimilarMobileCardClick 同採三層策略：
-        _filteredVideos miss 後先查 _videos.findIndex（命中保完整 metadata + path），
-        且仍保留 _similarLastDrilledItem snapshot 當孤兒列 fallback（Codex P2）。
-        否則 mobile tier2 的完整 metadata 會在關閉 similar mode 時被重新降級。
-        """
-        js = self._similar_js()
-        body = self._extract_function_body(js, r'async\s+closeSimilarMode\s*\(')
-        assert body is not None, \
-            "state-similar.js 找不到 closeSimilarMode 函式宣告"
-        assert re.search(r'_videos\s*\.\s*findIndex', body), (
-            "closeSimilarMode fallback 未含 '_videos.findIndex'（tier 2）。\n"
-            "退場時 _filteredVideos miss 直接降級成 5 欄 snapshot，\n"
-            "mobile tier2 點擊時的完整 metadata + path 會在關閉 similar mode 時被重新降級。"
-        )
-        assert '_similarLastDrilledItem' in body, (
-            "closeSimilarMode fallback 未保留 '_similarLastDrilledItem' snapshot（孤兒列 fallback）。\n"
-            "_videos 也 miss（孤兒列 / demo）時需 snapshot 兜底，不可移除。"
-        )
-
-    def test_mobile_silent_switch_three_tier(self):
-        """_mobileSilentSwitch 函式體必須採 3-tier silent-switch（恢復退役的 onSimilarMobileCardClick 守衛之 contract）：
-        tier1 _silentSwitchLightboxByNumber（_filteredVideos 命中 → _setLightboxIndex）、
-        tier2 _videos.findIndex（庫內 standalone，setSimilarExitVideo）、
-        tier3 snapshot（_mobileLastDrilledItem 孤兒列 fallback）、
-        且 tier2/3 收尾呼叫 _refreshLbFullBlurUp（blur-up reset，否則 overlay opacity:0 卡死）。
-        """
-        js = self._similar_js()
-        # 錨定方法宣告（行首縮排 + 名稱，非 this._mobileSilentSwitch( 呼叫處）
-        body = self._extract_function_body(js, r'\n\s*_mobileSilentSwitch\s*\(')
-        assert body is not None, \
-            "state-similar.js 找不到 _mobileSilentSwitch 函式宣告"
-        # tier 1: _filteredVideos 命中走 _silentSwitchLightboxByNumber
-        assert '_silentSwitchLightboxByNumber' in body, (
-            "_mobileSilentSwitch 缺 tier1 '_silentSwitchLightboxByNumber'（_filteredVideos 命中路徑）。"
-        )
-        # tier 2: _videos.findIndex 撈庫內被 filter 排除片的完整 metadata
-        assert re.search(r'_videos\s*\.\s*findIndex', body), (
-            "_mobileSilentSwitch 缺 tier2 '_videos.findIndex'（庫內 standalone metadata 撈回）。"
-        )
-        # tier 2/3 設 similarExitVideo standalone 旗標
-        assert re.search(r'similarExitVideo\s*=', body), (
-            "_mobileSilentSwitch 缺 'similarExitVideo =' 指派（tier2/3 standalone 旗標，close 不 fly-back / prev-next 禁用）。"
-        )
-        # tier 3: snapshot 兜底（孤兒列 / demo）
-        assert '_mobileLastDrilledItem' in body, (
-            "_mobileSilentSwitch 缺 '_mobileLastDrilledItem' snapshot（tier3 孤兒列 fallback）。"
-        )
-        # tier 2/3 收尾：blur-up reset
-        assert '_refreshLbFullBlurUp' in body, (
-            "_mobileSilentSwitch 缺 '_refreshLbFullBlurUp' 呼叫（tier2/3 blur-up reset，否則 overlay opacity:0 卡死）。"
-        )
 
 
 # ─── 90c-T5: external_manager switch-mode destructive confirm frontend guards ─
@@ -10018,311 +8246,6 @@ class TestExternalManagerSwitchModeGuard:
 SETTINGS_CSS = Path(__file__).parent.parent.parent / "web" / "static" / "css" / "pages" / "settings.css"
 
 
-class TestServerModeToggleGuard:
-    """80a-T3: settings.html + state-config.js server-mode toggle/banner 靜態守衛。
-
-    每個 assertion 都是 mutation-sensitive：刪除對應實作即 RED。
-    使用 BeautifulSoup DOM 解析（attribute 順序無關）+ substring 雙重策略，
-    鏡照既有 settings 守衛慣例（SETTINGS_HTML / SETTINGS_CONFIG_JS path 常數）。
-    """
-
-    def _html(self):
-        return SETTINGS_HTML.read_text(encoding="utf-8")
-
-    def _js(self):
-        return SETTINGS_CONFIG_JS.read_text(encoding="utf-8")
-
-    # ── HTML guards ────────────────────────────────────────────────────────────
-
-    def test_settings_root_has_data_lan_ip(self):
-        """#settings-components root div 含 data-lan-ip 屬性（傳 lanIp 到 Alpine）。
-        移除此屬性 → lanIp 永遠空字串，URL 顯示錯誤。"""
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._html(), "html.parser")
-        root = soup.find(id="settings-components")
-        assert root is not None, "settings.html 找不到 #settings-components"
-        assert root.has_attr("data-lan-ip"), \
-            "#settings-components 缺少 data-lan-ip 屬性（80a-T3 lanIp 傳入點）"
-
-    def test_settings_server_mode_segmented_in_header(self):
-        """.settings-server-mode 在標題左 cluster（.settings-header-left）內，是 <h4> 的
-        cluster-sibling，且含 .settings-sources-segmented + 2 個 button
-        （requestServerModeChange(false)/requestServerModeChange(true)，T4 改為確認 modal）。
-
-        81b-T1（CD-1）：膠囊從 .settings-header-actions 搬到 .settings-header-left。
-        mutation：膠囊搬回 .settings-header-actions → 「不在 actions」斷言 RED；
-        刪 .settings-header-left wrapper → 「在 header-left」斷言 RED。"""
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._html(), "html.parser")
-        wrapper = soup.find(class_="settings-server-mode")
-        assert wrapper is not None, \
-            "settings.html 缺少 .settings-server-mode wrapper（80a-T3 膠囊容器）"
-        segmented = wrapper.find(class_="settings-sources-segmented")
-        assert segmented is not None, \
-            ".settings-server-mode 缺少 .settings-sources-segmented（pill 複用）"
-        buttons = segmented.find_all("button")
-        assert len(buttons) == 2, \
-            f".settings-sources-segmented 應有 2 個 button，實際 {len(buttons)} 個"
-        click_attrs = [b.get("@click", "") for b in buttons]
-        assert any("requestServerModeChange(false)" in a for a in click_attrs), \
-            "segmented 缺少 @click=\"requestServerModeChange(false)\" button（單機態）"
-        assert any("requestServerModeChange(true)" in a for a in click_attrs), \
-            "segmented 缺少 @click=\"requestServerModeChange(true)\" button（伺服器態）"
-        # 81b-T1（CD-1）：位置斷言 — 膠囊在標題左 cluster，h4 sibling，搬離 actions
-        left = soup.find(class_="settings-header-left")
-        assert left is not None, \
-            "settings.html 缺少 .settings-header-left（81b-T1 標題左 cluster wrapper）"
-        assert left.find(class_="settings-server-mode") is not None, \
-            ".settings-server-mode 不在 .settings-header-left 內（CD-1 膠囊應在標題左 cluster）"
-        assert left.find("h4") is not None, \
-            "<h4> 不在 .settings-header-left 內（膠囊應與 h4 同 cluster）"
-        actions = soup.find(class_="settings-header-actions")
-        assert actions is None or actions.find(class_="settings-server-mode") is None, \
-            ".settings-server-mode 仍在 .settings-header-actions 內（CD-1 應已搬離至標題左 cluster）"
-
-    def test_settings_server_info_banner_xshow_xcloak(self):
-        """恰好一個 .settings-server-inline 元素，帶 x-show=\"serverMode\" 和 x-cloak（防 FOUC）。
-        移除 x-cloak → Alpine boot 前裸閃；移除 x-show → 區塊恆顯。
-        81b-T1：原獨立第二排 .settings-server-info 收進 header 同排 .settings-server-inline。"""
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._html(), "html.parser")
-        banners = soup.find_all(class_="settings-server-inline")
-        assert len(banners) == 1, \
-            f"settings.html .settings-server-inline 應恰好 1 個，實際 {len(banners)} 個"
-        banner = banners[0]
-        assert banner.get("x-show") == "serverMode", \
-            f".settings-server-inline x-show 應為 'serverMode'，實際: {banner.get('x-show')!r}"
-        assert banner.has_attr("x-cloak"), \
-            ".settings-server-inline 缺少 x-cloak（Alpine boot 前會 FOUC）"
-
-    def test_settings_server_info_warning_key(self):
-        """橫條內含 settings.server_info.warning i18n key 引用（警語行）。
-        移除此 key → 警語靜默，用戶不知安全風險。"""
-        html = self._html()
-        assert "settings.server_info.warning" in html, \
-            "settings.html 缺少 settings.server_info.warning i18n key 引用（警語行）"
-
-    def test_settings_server_info_copy_button(self):
-        """橫條內含 copyServerUrl() 呼叫（複製鈕 @click）。
-        移除 → 複製功能斷掉，用戶無法複製 URL。"""
-        html = self._html()
-        assert "copyServerUrl()" in html, \
-            "settings.html 缺少 copyServerUrl() 呼叫（複製鈕 @click）"
-
-    def test_settings_server_copy_is_clipboard_icon_with_aria(self):
-        """copy 鈕為 icon-only（內含 <i class="bi bi-clipboard">）+ a11y 標籤
-        （:aria-label 與 :title 皆引用 settings.server_info.copy）。
-
-        81b-T1（CD-4 + #8）：copy 由文字「複製」改 bi-clipboard icon-only，
-        i18n key 轉作 aria-label/title 提供無障礙標籤。
-        mutation：把 <i class="bi bi-clipboard"> 換回文字 → icon 斷言 RED；
-        移除 :aria-label 或改引用別 key → a11y 斷言 RED。"""
-        from bs4 import BeautifulSoup
-        soup = BeautifulSoup(self._html(), "html.parser")
-        btn = soup.find(class_="settings-server-copy-btn")
-        assert btn is not None, \
-            "settings.html 缺少 .settings-server-copy-btn（81b-T1 複製鈕）"
-        assert btn.find("i", class_="bi-clipboard") is not None, \
-            ".settings-server-copy-btn 缺少 <i class=\"bi bi-clipboard\"> icon（CD-4 icon-only）"
-        aria = btn.get(":aria-label") or btn.get("aria-label")
-        assert aria is not None and "settings.server_info.copy" in aria, \
-            f"copy 鈕 aria-label 應引用 settings.server_info.copy，實際: {aria!r}（#8 a11y 標籤）"
-        title = btn.get(":title") or btn.get("title")
-        assert title is not None and "settings.server_info.copy" in title, \
-            f"copy 鈕 title 應引用 settings.server_info.copy，實際: {title!r}（CD-4 hover 提示）"
-
-    def test_settings_amber_active_scoped_to_server_mode(self):
-        """琥珀 active（[data-mode="server"].is-on）規則 scope 緊收到
-        .settings-server-mode 且 body 用 --color-warning。
-
-        81b-T2（CD-6）：琥珀只套 server 膠囊，不可外溢到來源卡膠囊。
-        mutation：把任一條琥珀 server 規則的 scope 從 .settings-server-mode 拿掉
-        （污染來源卡）→ 該條成「未 scope 的琥珀規則」→ RED；改用非 --color-warning 顏色
-        → 找不到琥珀規則 → RED。
-
-        關鍵（teeth）：不是「存在一條有 scope 的規則就放行」（base + dim 兩條，拔一條
-        另一條仍在會假綠），而是「**每一條** [data-mode=\"server\"].is-on 琥珀規則都必須
-        scope 在 .settings-server-mode」——即不存在任何未 scope 的琥珀 server 規則。
-        先剝 /* ... */ 註解，防註解內 .settings-server-mode 字面混入 selector 比對。"""
-        css = SETTINGS_CSS.read_text(encoding="utf-8")
-        css = re.sub(r"/\*.*?\*/", "", css, flags=re.DOTALL)
-        # 列舉所有 `selector { body }` 規則塊（巢狀無關，settings.css 為扁平規則）
-        amber_rules = []
-        for m in re.finditer(r"([^{}]+)\{([^{}]*)\}", css):
-            selector, body = m.group(1), m.group(2)
-            if '[data-mode="server"]' in selector and ".is-on" in selector \
-                    and "--color-warning" in body:
-                amber_rules.append(selector.strip())
-        assert amber_rules, \
-            "settings.css 缺少 [data-mode=\"server\"].is-on + --color-warning 的琥珀 active 規則（CD-6）"
-        # 每一條琥珀 server 規則都必須 scope 在 .settings-server-mode（無任一條外溢）
-        unscoped = [s for s in amber_rules if ".settings-server-mode" not in s]
-        assert not unscoped, \
-            f"琥珀 active 規則未全部 scope 在 .settings-server-mode（CD-6 防外溢來源卡）: {unscoped!r}"
-
-    def test_no_clipboard_emoji_in_copy_context(self):
-        """web/ 下所有 .html/.js/.css 不得出現 📋 字面（含 design-system <pre> 範例字串）。
-
-        81b-T4：複製鈕統一 bi-clipboard icon，移除所有 📋 emoji。
-        純字串 in-content 掃描（非 bs4 text-only），確保 escaped <pre> 範例也掃到。
-        eslint/stylelint 不掃 .html 模板，故 pytest 字串掃描是正確路由。
-        mutation：任何人把 📋 寫回任一複製鈕（含教學範例）→ RED。"""
-        web_root = Path(__file__).parent.parent.parent / "web"
-        offenders = []
-        for ext in ("*.html", "*.js", "*.css"):
-            for f in web_root.rglob(ext):
-                if "\U0001F4CB" in f.read_text(encoding="utf-8"):
-                    offenders.append(str(f.relative_to(web_root)))
-        assert not offenders, \
-            f"web/ 下仍殘留 📋 emoji（應統一用 bi-clipboard icon）: {offenders}"
-
-    def test_settings_server_info_no_lan_ip_key(self):
-        """橫條含 settings.server_info.no_lan_ip i18n key（lanIp 為空時提示）。
-        移除 → lanIp=None 時橫條空白，用戶看不到說明。"""
-        html = self._html()
-        assert "settings.server_info.no_lan_ip" in html, \
-            "settings.html 缺少 settings.server_info.no_lan_ip i18n key（lanIp 空白提示）"
-
-    def test_settings_server_info_distinguishes_listener_down_from_no_ip(self):
-        """Codex P2：banner 須區分「listener 未啟動」(lanIp 在但 lanPort 缺) 與「取不到 IP」。
-        listener_down 分支須 lanIp-gated（`!serverUrl() && lanIp`），no_lan_ip 須 `!serverUrl() && !lanIp`。
-        若兩者合併回單一 `!serverUrl()` → autostart 失敗時誤顯「取不到 IP」誤導排查 → 此測 RED。"""
-        html = self._html()
-        assert "settings.server_info.listener_down" in html, \
-            "settings.html 缺少 settings.server_info.listener_down（listener 未啟動的專屬訊息）"
-        assert 'x-if="!serverUrl() && lanIp"' in html, \
-            "listener_down 分支須 gate 在 '!serverUrl() && lanIp'（lanIp 在但無 port）"
-        # Codex P2：抓不到 IP 但 port 已知 → 顯示帶 port 的訊息（用戶才湊得出 IP:port）；
-        # 連 port 都沒有才用純 no_lan_ip。兩支須各自 lanPort-gated。
-        assert "settings.server_info.no_lan_ip_with_port" in html, \
-            "settings.html 缺少 no_lan_ip_with_port（IP 抓不到但 lanPort 已知時顯示 port）"
-        assert 'x-if="!serverUrl() && !lanIp && lanPort"' in html, \
-            "no_lan_ip_with_port 分支須 gate 在 '!serverUrl() && !lanIp && lanPort'"
-        assert 'x-if="!serverUrl() && !lanIp && !lanPort"' in html, \
-            "no_lan_ip 分支須 gate 在 '!serverUrl() && !lanIp && !lanPort'（連 port 都沒有）"
-
-    # ── JS guards ──────────────────────────────────────────────────────────────
-
-    def test_state_config_server_mode_put_endpoint(self):
-        """state-config.js 含 PUT /api/config/general/server_mode endpoint。
-        移除或改路徑 → 後端收不到，config 不持久。"""
-        js = self._js()
-        assert "/api/config/general/server_mode" in js, \
-            "state-config.js 缺少 '/api/config/general/server_mode' PUT endpoint"
-
-    def test_state_config_server_url_uses_lan_port(self):
-        """serverUrl() 使用後端提供的 this.lanPort（非 window.location.port）。
-        T6c 起 dual-listener 後 LAN port ≠ 桌面 local_port，必須用後端回傳值。
-        回退至 window.location.port → 遠端裝置連到桌面 port 不是 LAN port。"""
-        js = self._js()
-        assert "this.lanPort" in js, \
-            "state-config.js serverUrl() 缺少 this.lanPort（80a-T6c：用後端 lan_port）"
-        assert "window.location.port" not in js, \
-            "state-config.js serverUrl() 不應再使用 window.location.port（已改用 lanPort）"
-
-    def test_state_config_lan_port_state_exists(self):
-        """state 含 lanPort: null（80a-T6c 新增 state，reload 後補 GET lan-port）。
-        移除 → serverUrl() 永遠 null，URL 橫條不顯示。"""
-        js = self._js()
-        assert "lanPort: null" in js, \
-            "state-config.js 缺少 'lanPort: null' state（80a-T6c）"
-
-    def test_state_config_set_server_mode_reads_lan_port(self):
-        """setServerMode() 成功分支讀 result.lan_port（後端回傳 LAN port）。
-        移除 → 切換成功後 lanPort 不更新，URL 橫條顯示舊值或 null。"""
-        js = self._js()
-        assert "result.lan_port" in js, \
-            "state-config.js setServerMode() 缺少 'result.lan_port'（80a-T6c：讀後端 lan_port）"
-
-    def test_state_config_set_server_mode_reads_lan_ip(self):
-        """setServerMode() 成功分支讀 result.lan_ip 並更新 this.lanIp。
-        移除 → 切換成功後 lanIp 不更新，serverUrl() 用舊值（或空字串 → null URL）。"""
-        js = self._js()
-        assert "result.lan_ip" in js, \
-            "state-config.js setServerMode() 缺少 'result.lan_ip'（gated get_lan_ip：由後端 toggle 回應提供）"
-        assert "this.lanIp = result.lan_ip" in js, \
-            "state-config.js setServerMode() 缺少 'this.lanIp = result.lan_ip'（lanIp 未從 toggle 回應更新）"
-
-    def test_state_config_load_config_fetches_lan_port(self):
-        """loadConfig() serverMode=true 時 GET /api/config/general/lan-port 補 lanPort。
-        移除 → reload 後 lanPort=null，URL 橫條消失（需重新 toggle 才恢復）。"""
-        js = self._js()
-        assert "/api/config/general/lan-port" in js, \
-            "state-config.js loadConfig() 缺少 '/api/config/general/lan-port' GET（80a-T6c）"
-
-    def test_state_config_load_config_reads_lan_ip_from_lan_port_endpoint(self):
-        """loadConfig() GET lan-port 分支讀 j.lan_ip 並更新 this.lanIp。
-        移除 → reload 後 lanIp 不更新（gated context：server_mode=true 時頁面 lan_ip 已空）。"""
-        js = self._js()
-        assert "j.lan_ip" in js, \
-            "state-config.js loadConfig() 缺少 'j.lan_ip'（lan-port GET 須同步更新 lanIp）"
-        assert "this.lanIp = j.lan_ip" in js, \
-            "state-config.js loadConfig() 缺少 'this.lanIp = j.lan_ip'（lanIp 未從 lan-port 回應更新）"
-
-    def test_state_config_reads_server_mode_with_nullish_coalesce(self):
-        """loadConfig() 用 ?? false 讀 config.general?.server_mode（CD#3 慣例）。
-        改成 || false → false 值會被吞（語意等同但守慣例）。"""
-        js = self._js()
-        assert "config.general?.server_mode ?? false" in js, \
-            "state-config.js loadConfig() 缺少 'config.general?.server_mode ?? false'"
-
-    def test_state_config_set_server_mode_failure_direction_aware(self):
-        """setServerMode() 失敗分支使用方向感知 toast（val ? toggle_failed : disable_failed）。
-
-        移除三元表達式或合併成同一 key：
-          - enable 失敗顯示「無法啟動」✓，disable 失敗仍顯「無法啟動」✗（語意錯誤）。
-        兩個 key 都必須存在，且須以 `val ?` ternary 選擇（mutation-sensitive）。
-        """
-        js = self._js()
-        assert "settings.server_info.disable_failed" in js, (
-            "state-config.js setServerMode() 失敗分支缺少 'settings.server_info.disable_failed'（80b：方向感知 toast）"
-        )
-        assert "settings.server_info.toggle_failed" in js, (
-            "state-config.js setServerMode() 失敗分支應保留 'settings.server_info.toggle_failed'（enable 失敗路徑）"
-        )
-        # Ternary pattern: val ? 'toggle_failed' : 'disable_failed'（或反序帶 ! 亦合法，
-        # 但實作固定 val ? toggle_failed : disable_failed，故字串比對足以 mutation-catch）
-        assert "val ? 'settings.server_info.toggle_failed' : 'settings.server_info.disable_failed'" in js, (
-            "state-config.js setServerMode() 失敗分支須以 ternary 按 val 選擇 key（80b：方向感知）"
-        )
-        # 遠端被拒（loopback 守衛）須給專屬 remote_only 訊息，而非通用「請稍後再試」
-        assert "remote_forbidden" in js and "settings.server_info.remote_only" in js, (
-            "state-config.js setServerMode() 失敗分支須對 reason==='remote_forbidden' 顯示 "
-            "'settings.server_info.remote_only'（遠端切換給專屬訊息，非『請稍後再試』）"
-        )
-
-    def test_state_config_set_server_mode_sends_boolean(self):
-        """setServerMode() body: JSON.stringify({ value: !!val })，確保送 boolean 不送 string。
-        T1 嚴格 bool gate，送字串會 400。"""
-        js = self._js()
-        assert "value: !!val" in js, \
-            "state-config.js setServerMode() 缺少 'value: !!val'（必須送 boolean，不可送字串）"
-
-    def test_set_server_mode_lan_ip_nullish_uses_null_not_stale(self):
-        """P2-2: setServerMode() enable 成功分支用 `result.lan_ip ?? null`，不用 `?? this.lanIp`。
-        用 `?? this.lanIp` → 後端返 null（IP 偵測失敗）時 banner 仍顯示舊 IP，指向失效 URL。
-        必須用 `?? null` 讓 serverUrl() 返 null，正確顯示 no_lan_ip/listener_down 提示。"""
-        js = self._js()
-        assert "result.lan_ip ?? null" in js, (
-            "state-config.js setServerMode() 須用 'result.lan_ip ?? null'（P2-2）；"
-            "不可用 '?? this.lanIp'（保留舊 IP 會在 IP 偵測失敗時顯示死連結）"
-        )
-        assert "result.lan_ip ?? this.lanIp" not in js, (
-            "state-config.js setServerMode() 不得用 'result.lan_ip ?? this.lanIp'（P2-2 stale IP bug）"
-        )
-
-    def test_load_config_lan_ip_nullish_uses_null_not_stale(self):
-        """P2-2: loadConfig() GET lan-port 分支用 `j.lan_ip ?? null`，不用 `?? this.lanIp`。
-        用 `?? this.lanIp` → 重載後 lan_ip=null（偵測失敗）時 banner 仍顯示舊 IP。
-        必須用 `?? null` 讓 serverUrl() 返 null 觸發正確的 no_lan_ip 提示路徑。"""
-        js = self._js()
-        assert "j.lan_ip ?? null" in js, (
-            "state-config.js loadConfig() 須用 'j.lan_ip ?? null'（P2-2）；"
-            "不可用 '?? this.lanIp'（保留舊 IP 會在 IP 偵測失敗時顯示死連結）"
-        )
-        assert "j.lan_ip ?? this.lanIp" not in js, (
-            "state-config.js loadConfig() 不得用 'j.lan_ip ?? this.lanIp'（P2-2 stale IP bug）"
-        )
 
 
 class TestMobileToolbarToggle:
@@ -10602,322 +8525,12 @@ T11_SHOWCASE_CSS      = PROJECT_ROOT / "web" / "static" / "css" / "pages" / "sho
 T11_SEARCH_CSS        = PROJECT_ROOT / "web" / "static" / "css" / "pages" / "search.css"
 
 
-class TestPosterCropThresholdAlignment:
-    """US-10 / CD-10：posterCrop 門檻（JS）與燈箱封面貼合 / poster grid 斷點（CSS）對齊 899。"""
-
-    POSTER_CROP_CONST = "POSTER_CROP_MAX_W"
-
-    # ---- helpers ----
-    def _const_value(self) -> int:
-        """從 shared/breakpoints.js 抽 POSTER_CROP_MAX_W 常數值。"""
-        content = T11_BREAKPOINTS_JS.read_text(encoding="utf-8")
-        m = re.search(
-            rf"export\s+const\s+{self.POSTER_CROP_CONST}\s*=\s*(\d+)", content
-        )
-        assert m, f"breakpoints.js 缺少 `export const {self.POSTER_CROP_CONST} = <int>`"
-        return int(m.group(1))
-
-    def _js_poster_crop_threshold(self, path: Path, label: str) -> int:
-        """抽某 JS 檔的 posterCrop 門檻值。
-
-        共用常數方案：斷言檔案 import 了 POSTER_CROP_MAX_W 並用於 innerWidth 比較，
-        門檻值即常數值（單一真理來源在 breakpoints.js）。
-        """
-        content = path.read_text(encoding="utf-8")
-        # 1. import 共用常數（物理同源）
-        assert re.search(
-            rf"import\s*\{{[^}}]*\b{self.POSTER_CROP_CONST}\b[^}}]*\}}\s*from\s*"
-            r"['\"]@/shared/breakpoints\.js['\"]",
-            content,
-        ), f"{label} 必須 import {{ {self.POSTER_CROP_CONST} }} from '@/shared/breakpoints.js'"
-        # 2. 用於 innerWidth 門檻比較
-        assert re.search(
-            rf"window\.innerWidth\s*<=\s*{self.POSTER_CROP_CONST}\b", content
-        ), f"{label} 必須以 `window.innerWidth <= {self.POSTER_CROP_CONST}` 作 posterCrop 門檻"
-        # 3. 舊的 480 字面量門檻已消失（防只改一半）
-        assert not re.search(
-            r"window\.innerWidth\s*<=\s*480\b", content
-        ), f"{label} 仍殘留 `window.innerWidth <= 480`（posterCrop 門檻未擴）"
-        return self._const_value()
-
-    def _css_media_max_width_for_selector(
-        self, css: str, selector_substr: str, name: str
-    ) -> int:
-        """找含 selector_substr 的 @media 區塊，回傳其 max-width 值。"""
-        for m in re.finditer(r"@media\s*\(([^)]*max-width[^)]*)\)\s*\{", css):
-            start = m.end()
-            depth = 1
-            i = start
-            while i < len(css) and depth > 0:
-                if css[i] == "{":
-                    depth += 1
-                elif css[i] == "}":
-                    depth -= 1
-                i += 1
-            body = css[start : i - 1]
-            if selector_substr in body:
-                mw = re.search(r"max-width:\s*(\d+)px", m.group(1))
-                assert mw, f"{name} 的 @media 條件抽不到 max-width: {m.group(1)!r}"
-                return int(mw.group(1))
-        raise AssertionError(f"{name} 找不到含 selector {selector_substr!r} 的 @media block")
-
-    def _poster_grid_max_width(self, css: str, grid_class: str, name: str) -> int:
-        """T10 poster-crop grid 斷點：回傳「481–899 4 欄直式右裁 poster」@media 的 max-width。
-
-        grid_class（.showcase-grid / .search-grid）在多個 @media 出現（≤480、481–899、
-        1100–1499 桌面等都可能是 repeat(4)）。poster-crop 斷點唯一特徵 = 下界 481px：
-        鎖 `(min-width: 481px) and (max-width: Npx)` 且 grid 為 repeat(4) 的那塊，回傳 N。
-        """
-        for m in re.finditer(r"@media\s*([^{]*?)\s*\{", css):
-            cond = m.group(1)
-            mn = re.search(r"min-width:\s*481px", cond)
-            mw = re.search(r"max-width:\s*(\d+)px", cond)
-            if not (mn and mw):
-                continue
-            start = m.end()
-            depth, i = 1, start
-            while i < len(css) and depth > 0:
-                if css[i] == "{":
-                    depth += 1
-                elif css[i] == "}":
-                    depth -= 1
-                i += 1
-            body = css[start : i - 1]
-            if re.search(
-                rf"\.{re.escape(grid_class)}\b[^{{}}]*\{{[^}}]*repeat\(\s*4\b", body, re.DOTALL
-            ):
-                return int(mw.group(1))
-        raise AssertionError(
-            f"{name} 找不到 (min-width:481px)+(max-width)+.{grid_class} repeat(4) poster grid @media block"
-        )
-
-    # ---- 常數定義 ----
-    def test_breakpoint_const_is_899(self):
-        """shared/breakpoints.js 匯出 POSTER_CROP_MAX_W = 899。"""
-        assert self._const_value() == 899, "POSTER_CROP_MAX_W 不為 899"
-
-    # ---- JS 門檻 ----
-    def test_showcase_js_threshold_899(self):
-        """state-lightbox.js posterCrop 門檻 == 899（import 共用常數 + innerWidth 比較 + 無 480 殘留）。"""
-        assert self._js_poster_crop_threshold(
-            T11_STATE_LIGHTBOX_JS, "state-lightbox.js"
-        ) == 899
-
-    def test_search_js_threshold_899(self):
-        """grid-mode.js（search）posterCrop 門檻 == 899（同上）。"""
-        assert self._js_poster_crop_threshold(
-            T11_GRID_MODE_JS, "grid-mode.js"
-        ) == 899
-
-    def test_parity_both_js_thresholds_equal(self):
-        """CD-11：兩頁 JS 門檻值彼此相等（防只改一頁）。"""
-        sc = self._js_poster_crop_threshold(T11_STATE_LIGHTBOX_JS, "state-lightbox.js")
-        se = self._js_poster_crop_threshold(T11_GRID_MODE_JS, "grid-mode.js")
-        assert sc == se, f"兩頁 posterCrop 門檻不一致：showcase={sc} search={se}"
-
-    # ---- CSS 燈箱封面貼合 ----
-    def test_showcase_lightbox_fit_covers_899(self):
-        """showcase.css 燈箱封面貼合：83b-T2 後由 modal-hug 規則（.lightbox-content .lightbox-cover.has-cover img）
-        無條件提供 width:100%（不再依賴 @media max-width:899px T8 block）。
-        守衛改為確認 modal-hug img 規則存在且含 width:100%。
-        """
-        css = T11_SHOWCASE_CSS.read_text(encoding="utf-8")
-        # 83b-T2 移除 T8 block，modal-hug 接管（無 @media gate）
-        assert ".lightbox-content .lightbox-cover.has-cover img" in css, (
-            "showcase.css modal-hug img 規則缺失（83b-T2 後應由此規則提供 width:100%）"
-        )
-        # 確認 img 規則內含 width:100%（strip comments 後查）
-        css_no_comments = re.sub(r'/\*.*?\*/', '', css, flags=re.DOTALL)
-        m = re.search(r'\.lightbox-content\s+\.lightbox-cover\.has-cover\s+img\s*\{([^}]+)\}', css_no_comments)
-        assert m, "showcase.css modal-hug img block 找不到"
-        assert "width: 100%" in m.group(1), "modal-hug img block 缺 width: 100%"
-
-    def test_search_lightbox_fit_covers_899(self):
-        """search.css 燈箱封面貼合（.search-container .lightbox-cover.has-cover）@media == max-width: 899px。"""
-        css = T11_SEARCH_CSS.read_text(encoding="utf-8")
-        mw = self._css_media_max_width_for_selector(
-            css, ".search-container .lightbox-cover.has-cover", "search.css 燈箱貼合"
-        )
-        assert mw == 899, f"search.css 燈箱貼合 @media max-width={mw}px，應為 899"
-
-    # ---- T10 poster grid 斷點（參考；三位一體比對）----
-    def test_showcase_poster_grid_breakpoint_899(self):
-        """showcase.css poster-crop grid 斷點（.showcase-grid 4 欄）== max-width: 899px（T10）。"""
-        css = T11_SHOWCASE_CSS.read_text(encoding="utf-8")
-        mw = self._poster_grid_max_width(css, "showcase-grid", "showcase.css")
-        assert mw == 899, f"showcase.css poster grid @media max-width={mw}px，應為 899"
-
-    def test_search_poster_grid_breakpoint_899(self):
-        """search.css poster-crop grid 斷點（.search-grid 4 欄）== max-width: 899px（T10）。"""
-        css = T11_SEARCH_CSS.read_text(encoding="utf-8")
-        mw = self._poster_grid_max_width(css, "search-grid", "search.css")
-        assert mw == 899, f"search.css poster grid @media max-width={mw}px，應為 899"
-
-    # ---- 核心：跨 6 值單一對齊斷言 ----
-    def test_all_thresholds_aligned_899(self):
-        """核心對齊守衛：2 JS 門檻 + search CSS 燈箱貼合 + 2 CSS poster grid 全部相等且 == 899。
-
-        83b-T2：showcase.css T8 block（@media max-width:899px .lightbox-cover:has(.lb-full)）已移除；
-        showcase 燈箱貼合改由 modal-hug 無條件提供。核心對齊守衛去掉 showcase-lightbox-fit 維度，
-        保留其他 5 值（2 JS + 1 search CSS lightbox-fit + 2 poster grid）。
-        """
-        showcase_css = T11_SHOWCASE_CSS.read_text(encoding="utf-8")
-        search_css = T11_SEARCH_CSS.read_text(encoding="utf-8")
-        values = {
-            "js:showcase": self._js_poster_crop_threshold(
-                T11_STATE_LIGHTBOX_JS, "state-lightbox.js"
-            ),
-            "js:search": self._js_poster_crop_threshold(
-                T11_GRID_MODE_JS, "grid-mode.js"
-            ),
-            "css:search-lightbox-fit": self._css_media_max_width_for_selector(
-                search_css, ".search-container .lightbox-cover.has-cover", "search.css 燈箱貼合"
-            ),
-            "css:showcase-poster-grid": self._poster_grid_max_width(
-                showcase_css, "showcase-grid", "showcase.css"
-            ),
-            "css:search-poster-grid": self._poster_grid_max_width(
-                search_css, "search-grid", "search.css"
-            ),
-        }
-        assert all(v == 899 for v in values.values()), (
-            f"posterCrop 門檻 ↔ 燈箱貼合 ↔ poster grid 斷點未全對齊 899：{values}"
-        )
 
 
 # ---------------------------------------------------------------------------
 # feature/82 T4: Settings closeAction select guard
 # ---------------------------------------------------------------------------
 
-class TestSettingsCloseActionSelect:
-    """Guards for the closeAction select row in settings.html (feature/82 T4)."""
-
-    SETTINGS_HTML = (
-        Path(__file__).parents[2] / "web" / "templates" / "settings.html"
-    )
-    ZH_TW_JSON = Path(__file__).parents[2] / "locales" / "zh_TW.json"
-    SETTINGS_JS = (
-        Path(__file__).parents[2]
-        / "web" / "static" / "js" / "pages" / "settings" / "state-config.js"
-    )
-
-    def _settings_html(self) -> str:
-        return self.SETTINGS_HTML.read_text(encoding="utf-8")
-
-    def _zh_tw(self) -> dict:
-        import json
-        return json.loads(self.ZH_TW_JSON.read_text(encoding="utf-8"))
-
-    def test_close_action_select_inside_jinja_gate(self):
-        """#closeAction select element must be inside {% if is_windows_desktop %} … {% endif %}."""
-        import re
-        content = self._settings_html()
-
-        # Extract all {% if is_windows_desktop %} ... {% endif %} blocks
-        gate_pattern = re.compile(
-            r'\{%-?\s*if\s+is_windows_desktop\s*-?%\}(.*?)\{%-?\s*endif\s*-?%\}',
-            re.DOTALL,
-        )
-        gates = gate_pattern.findall(content)
-        assert gates, "No {% if is_windows_desktop %} block found in settings.html"
-
-        # closeAction select must appear inside at least one gate block
-        found_in_gate = any('id="closeAction"' in block for block in gates)
-        assert found_in_gate, (
-            '#closeAction select must be inside {% if is_windows_desktop %} block, '
-            'not outside the gate'
-        )
-
-    def test_close_action_select_not_outside_gate(self):
-        """#closeAction select must NOT appear outside the is_windows_desktop gate."""
-        import re
-        content = self._settings_html()
-
-        # Remove all gated blocks
-        gate_pattern = re.compile(
-            r'\{%-?\s*if\s+is_windows_desktop\s*-?%\}.*?\{%-?\s*endif\s*-?%\}',
-            re.DOTALL,
-        )
-        stripped = gate_pattern.sub('', content)
-        assert 'id="closeAction"' not in stripped, (
-            '#closeAction found outside {% if is_windows_desktop %} block — '
-            'non-desktop clients would receive the DOM element'
-        )
-
-    def test_close_action_select_has_x_model(self):
-        """The #closeAction select must have x-model='form.closeAction'."""
-        import re
-        content = self._settings_html()
-
-        # Find the select tag with id="closeAction"
-        m = re.search(r'<select\b[^>]*\bid="closeAction"[^>]*>', content, re.DOTALL)
-        assert m, 'select#closeAction not found in settings.html'
-        tag = m.group(0)
-        assert 'x-model="form.closeAction"' in tag, (
-            f'select#closeAction missing x-model="form.closeAction": {tag!r}'
-        )
-
-    def test_close_action_select_has_three_option_values(self):
-        """The #closeAction select must have options for ask, tray, and exit."""
-        import re
-        content = self._settings_html()
-
-        # Find the block from id="closeAction" up to closing </select>
-        m = re.search(
-            r'<select\b[^>]*\bid="closeAction"[^>]*>.*?</select>',
-            content, re.DOTALL,
-        )
-        assert m, 'select#closeAction not found in settings.html'
-        block = m.group(0)
-
-        for val in ("ask", "tray", "exit"):
-            assert f'value="{val}"' in block, (
-                f'select#closeAction missing option value="{val}": {block!r}'
-            )
-
-    def test_close_action_select_has_i18n_keys(self):
-        """Each option in the #closeAction select uses a t('settings.system.close_action_*') key."""
-        import re
-        content = self._settings_html()
-
-        m = re.search(
-            r'<select\b[^>]*\bid="closeAction"[^>]*>.*?</select>',
-            content, re.DOTALL,
-        )
-        assert m, 'select#closeAction not found in settings.html'
-        block = m.group(0)
-
-        for key in ("settings.system.close_action_ask",
-                    "settings.system.close_action_tray",
-                    "settings.system.close_action_exit"):
-            assert key in block, (
-                f'select#closeAction missing i18n key {key!r} in option text'
-            )
-
-    def test_zh_tw_has_close_action_label_key(self):
-        """zh_TW.json must contain settings.system.close_action_label."""
-        data = self._zh_tw()
-        system = data.get("settings", {}).get("system", {})
-        assert "close_action_label" in system, (
-            "zh_TW.json missing settings.system.close_action_label"
-        )
-        assert system["close_action_label"] == "關閉視窗時"
-
-    def test_zh_tw_has_close_action_option_keys(self):
-        """zh_TW.json must contain all 3 close_action option keys."""
-        data = self._zh_tw()
-        system = data.get("settings", {}).get("system", {})
-        expected = {
-            "close_action_ask": "每次詢問",
-            "close_action_tray": "最小化到系統匣",
-            "close_action_exit": "直接結束",
-        }
-        for key, value in expected.items():
-            assert key in system, f"zh_TW.json missing settings.system.{key}"
-            assert system[key] == value, (
-                f"settings.system.{key} value mismatch: "
-                f"expected {value!r}, got {system[key]!r}"
-            )
 
 
 # ---------------------------------------------------------------------------
@@ -11368,303 +8981,6 @@ _T2_BURST_PICKER_JS = (
 )
 
 
-class TestMobileSimilarPanelContractGuard:
-    """TASK-83b-T2: 行動相似面板（.similar-mobile-panel）13 條合約守衛。
-
-    鎖住 T1 建立的 CSS / HTML / JS 合約，防回歸。
-    每條 guard 均 mutation-detectable：刪除對應實作即 RED。
-    """
-
-    @staticmethod
-    def _html():
-        return _T2_SHOWCASE_HTML.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _css():
-        return _T2_SHOWCASE_CSS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _similar_js():
-        return _T2_SIMILAR_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _lightbox_js():
-        return _T2_LIGHTBOX_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _base_js():
-        return _T2_BASE_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _extract_function_body(js, func_pattern):
-        """大括號平衡法擷取函式體。"""
-        m = re.search(func_pattern, js)
-        if not m:
-            return None
-        start = m.start()
-        body_start = js.index('{', start)
-        depth = 0
-        for i, ch in enumerate(js[body_start:], body_start):
-            if ch == '{':
-                depth += 1
-            elif ch == '}':
-                depth -= 1
-                if depth == 0:
-                    return js[body_start:i + 1]
-        return None
-
-    # ── 1. CSS default-hidden + .show visible ──────────────────────────────
-
-    def test_mobile_panel_default_hidden(self):
-        """.similar-mobile-panel 存在於 HTML；CSS 含 default-hidden（opacity:0/visibility:hidden/
-        pointer-events:none）+ .show block（opacity:1/visibility:visible/pointer-events:auto）。"""
-        html = self._html()
-        assert 'class="similar-mobile-panel"' in html, \
-            "showcase.html 缺 .similar-mobile-panel div"
-        css = self._css()
-        # default-hidden block（strip CSS comment 後查，commented-out 行也應 RED）
-        m_panel = re.search(r'\.similar-mobile-panel\s*\{([^}]+)\}', css, re.DOTALL)
-        assert m_panel, "showcase.css 缺 .similar-mobile-panel default-hidden block"
-        block = re.sub(r'/\*.*?\*/', '', m_panel.group(1), flags=re.DOTALL)
-        assert "opacity: 0" in block, \
-            ".similar-mobile-panel block 缺 opacity: 0（FOUC 防護）"
-        assert "visibility: hidden" in block, \
-            ".similar-mobile-panel block 缺 visibility: hidden"
-        assert "pointer-events: none" in block, \
-            ".similar-mobile-panel block 缺 pointer-events: none"
-        # .show block
-        m_show = re.search(r'\.similar-mobile-panel\.show\s*\{([^}]+)\}', css, re.DOTALL)
-        assert m_show, "showcase.css 缺 .similar-mobile-panel.show block"
-        show_block = re.sub(r'/\*.*?\*/', '', m_show.group(1), flags=re.DOTALL)
-        assert "opacity: 1" in show_block, \
-            ".similar-mobile-panel.show 缺 opacity: 1"
-        assert "visibility: visible" in show_block, \
-            ".similar-mobile-panel.show 缺 visibility: visible"
-        assert "pointer-events: auto" in show_block, \
-            ".similar-mobile-panel.show 缺 pointer-events: auto"
-
-    # ── 2. Desktop safety net ───────────────────────────────────────────────
-
-    def test_mobile_panel_desktop_safety_net(self):
-        """showcase.css @media (min-width:960px) 內含 similar-mobile-panel + display:none（桌面安全網）。"""
-        css = self._css()
-        # 找含 similar-mobile-panel 的那個 @media (min-width:960px) block（可能有多個同斷點）
-        found = False
-        for m in re.finditer(r'@media\s*\(\s*min-width\s*:\s*960px\s*\)', css):
-            window = css[m.start():m.start() + 500]
-            if "similar-mobile-panel" in window and "display: none" in window:
-                found = True
-                break
-        assert found, (
-            "showcase.css 缺含 similar-mobile-panel + display:none 的 @media (min-width: 960px) block"
-            "（桌面安全網缺失，面板在桌面可能顯示）"
-        )
-
-    # ── 3. HTML x-trap ─────────────────────────────────────────────────────
-
-    def test_mobile_panel_has_x_trap(self):
-        """showcase.html .similar-mobile-panel block 含 x-trap.inert=\"similarModeMobileOpen\"。"""
-        html = self._html()
-        # 找 .similar-mobile-panel div 開始的 block
-        m = re.search(r'class="similar-mobile-panel"[^>]*>(.*?)</div>', html, re.DOTALL)
-        # 寬鬆：直接全文搜尋（similar-mobile-panel 唯一，不會誤中）
-        idx_panel = html.find('class="similar-mobile-panel"')
-        assert idx_panel != -1, "showcase.html 缺 similar-mobile-panel"
-        # x-trap 必須在 panel div 開啟標籤附近（同一 tag attribute）
-        panel_tag_end = html.index('>', idx_panel)
-        panel_opening_tag = html[idx_panel:panel_tag_end + 1]
-        assert 'x-trap.inert="similarModeMobileOpen"' in panel_opening_tag, \
-            "similar-mobile-panel div 開啟標籤缺 x-trap.inert=\"similarModeMobileOpen\""
-
-    # ── 4. Lightbox trap yields to mobile panel ─────────────────────────────
-
-    def test_mobile_panel_lightbox_trap_yields(self):
-        """showcase.html lightbox x-trap.inert 含 similarModeMobileOpen 條件（!similarModeMobileOpen）。
-        面板開時 trap 釋放給面板（防焦點被困在 lightbox）。
-        """
-        html = self._html()
-        # 錨定含 deleteVideoModalOpen 的那條（lightbox x-trap，T2 rewrite 後的錨點）
-        m = re.search(r'x-trap\.inert="([^"]*deleteVideoModalOpen[^"]*)"', html)
-        assert m, "showcase.html 缺含 deleteVideoModalOpen 的 x-trap.inert 行（lightbox trap）"
-        expr = m.group(1)
-        assert "similarModeMobileOpen" in expr, \
-            f"lightbox x-trap.inert 未含 similarModeMobileOpen: {expr!r}"
-        assert "!similarModeMobileOpen" in expr, \
-            f"lightbox x-trap.inert 缺 !similarModeMobileOpen（面板開時 trap 未釋放）: {expr!r}"
-
-    # ── 5. Burst card CSS ───────────────────────────────────────────────────
-
-    def test_mobile_burst_card_class_exists(self):
-        """.similar-mobile-burst-card block 存在；含 opacity:0、position:relative、transition:none。"""
-        css = self._css()
-        m = re.search(r'\.similar-mobile-burst-card\s*\{([^}]+)\}', css, re.DOTALL)
-        assert m, "showcase.css 缺 .similar-mobile-burst-card block（T1 burst card CSS 未被誤刪）"
-        block = re.sub(r'/\*.*?\*/', '', m.group(1), flags=re.DOTALL)
-        assert "opacity: 0" in block, \
-            ".similar-mobile-burst-card 缺 opacity: 0（GSAP-only，防 1-frame paint glitch）"
-        assert "position: relative" in block, \
-            ".similar-mobile-burst-card 缺 position: relative"
-        assert "transition: none" in block, \
-            ".similar-mobile-burst-card 缺 transition: none（GSAP-only）"
-
-    def test_mobile_burst_card_img_poster_crop(self):
-        """.similar-mobile-burst-card img block 含 aspect-ratio:var(--poster-crop-ratio)（單一真理、禁硬編碼）
-        + object-position:right center（右半裁切）。
-        恢復退役的 test_similar_mobile_card_img_has_poster_crop_ratio / _no_hardcoded_4_5 之 contract。
-        """
-        css = self._css()
-        m = re.search(r'\.similar-mobile-burst-card\s+img\s*\{([^}]+)\}', css, re.DOTALL)
-        assert m, "showcase.css 缺 .similar-mobile-burst-card img block"
-        block = re.sub(r'/\*.*?\*/', '', m.group(1), flags=re.DOTALL)
-        assert "aspect-ratio: var(--poster-crop-ratio)" in block, (
-            ".similar-mobile-burst-card img 缺 aspect-ratio: var(--poster-crop-ratio)（單一真理，禁硬編碼比例）"
-        )
-        assert "4/5" not in block, \
-            ".similar-mobile-burst-card img 不得含硬編碼 4/5（應用 var(--poster-crop-ratio)）"
-        assert "object-position: right center" in block, (
-            ".similar-mobile-burst-card img 缺 object-position: right center（右半裁切）"
-        )
-
-    # ── 6. Scrim blur token ─────────────────────────────────────────────────
-
-    def test_mobile_panel_scrim_blur_token(self):
-        """.similar-mobile-scrim 含 var(--fluent-blur)（不硬編碼 px）且含 -webkit-backdrop-filter。"""
-        css = self._css()
-        m = re.search(r'\.similar-mobile-scrim\s*\{([^}]+)\}', css, re.DOTALL)
-        assert m, "showcase.css 缺 .similar-mobile-scrim block"
-        block = m.group(1)
-        assert "var(--fluent-blur)" in block, \
-            ".similar-mobile-scrim backdrop-filter 必須用 var(--fluent-blur)（禁硬編碼 px）"
-        assert "-webkit-backdrop-filter" in block, \
-            ".similar-mobile-scrim 缺 -webkit-backdrop-filter（Safari fallback）"
-
-    # ── 7. Drill lock before await ──────────────────────────────────────────
-
-    def test_mobile_drill_lock_before_await(self):
-        """onMobileDrillClick 函式體在首個 await keyword 之前有 similarModeAnimating = true（lock-before-await）。"""
-        js = self._similar_js()
-        body = self._extract_function_body(js, r'async\s+onMobileDrillClick\s*\(')
-        assert body is not None, \
-            "state-similar.js 找不到 onMobileDrillClick 函式宣告"
-        # 去掉 // 單行 comment 和 /* */ 多行 comment，避免 "lock-before-await" 誤中
-        body_no_comments = re.sub(r'//[^\n]*', '', body)
-        body_no_comments = re.sub(r'/\*.*?\*/', '', body_no_comments, flags=re.DOTALL)
-        # 找首個 await 作為 JS 關鍵字（空白/符號前置，非 -await 形式）
-        m_await = re.search(r'(?<![A-Za-z0-9_$-])\bawait\b', body_no_comments)
-        assert m_await is not None, "onMobileDrillClick 函式體找不到 await keyword（非 async 路徑？）"
-        pre_await = body_no_comments[:m_await.start()]
-        assert "similarModeAnimating = true" in pre_await, (
-            "onMobileDrillClick 在首個 await keyword 之前缺 similarModeAnimating = true（lock-before-await 缺失，"
-            "連點空窗導致並發進入）"
-        )
-
-    # ── 8. closeMobilePanel does NOT call closeSimilarMode ──────────────────
-
-    def test_mobile_panel_no_call_desktop_closeSimilarMode(self):
-        """state-similar.js closeMobilePanel 函式體不含 closeSimilarMode() 呼叫（CD-4 / R3 防凍結）。
-        注：去掉 comment 後檢查（comment 中提及 closeSimilarMode 做對比說明是合法的）。
-        """
-        js = self._similar_js()
-        body = self._extract_function_body(js, r'closeMobilePanel\s*\(\s*\)\s*\{')
-        assert body is not None, \
-            "state-similar.js 找不到 closeMobilePanel 函式宣告"
-        # 去掉 comment，避免說明性文字（如 mirror closeSimilarMode）誤觸
-        body_no_comments = re.sub(r'//[^\n]*', '', body)
-        body_no_comments = re.sub(r'/\*.*?\*/', '', body_no_comments, flags=re.DOTALL)
-        assert "closeSimilarMode" not in body_no_comments, (
-            "closeMobilePanel 函式體含 closeSimilarMode() 呼叫（非 comment）！"
-            "CD-4/R3：桌面 close 在 similarCards={} 時 await playExit 永不 resolve → 凍結"
-        )
-
-    # ── 9. Burst card count = 6 ─────────────────────────────────────────────
-
-    def test_mobile_burst_card_count_6(self):
-        """state-similar.js _openMobilePanel 函式體含 slice(0, 6)（CD-6 固定 6 張）。"""
-        js = self._similar_js()
-        body = self._extract_function_body(js, r'async\s+_openMobilePanel\s*\(')
-        assert body is not None, \
-            "state-similar.js 找不到 _openMobilePanel 函式宣告"
-        assert "slice(0, 6)" in body, (
-            "_openMobilePanel 函式體缺 slice(0, 6)（CD-6：行動面板固定 6 張，不多不少）"
-        )
-
-    # ── 10. Uses own picker params ───────────────────────────────────────────
-
-    def test_mobile_uses_own_picker_params(self):
-        """state-similar.js 含 _MOBILE_PICKER_PARAMS local const 定義；
-        且程式碼（非 comment）中不含裸 _PICKER_PARAMS 字串（不直接引用 state-lightbox 的 private const）。
-        """
-        js = self._similar_js()
-        assert "_MOBILE_PICKER_PARAMS" in js, \
-            "state-similar.js 缺 _MOBILE_PICKER_PARAMS（行動面板專屬 picker params 未定義）"
-        # const 定義（不只是字串使用）
-        assert re.search(r'const\s+_MOBILE_PICKER_PARAMS', js), \
-            "state-similar.js 缺 const _MOBILE_PICKER_PARAMS 定義"
-        # 去掉 comment，再去掉 _MOBILE_PICKER_PARAMS，剩下若有 _PICKER_PARAMS 即裸引用
-        js_no_comments = re.sub(r'//[^\n]*', '', js)
-        js_no_comments = re.sub(r'/\*.*?\*/', '', js_no_comments, flags=re.DOTALL)
-        js_stripped = js_no_comments.replace("_MOBILE_PICKER_PARAMS", "")
-        assert "_PICKER_PARAMS" not in js_stripped, (
-            "state-similar.js 程式碼（非 comment）含裸 _PICKER_PARAMS 引用（state-lightbox private const，"
-            "直接引用在 stateSimilar 作用域會 ReferenceError）"
-        )
-
-    # ── 11. matchMedia 960 reset ─────────────────────────────────────────────
-
-    def test_mobile_panel_matchmedia_960(self):
-        """state-base.js 含 matchMedia + 960 + similarModeMobileOpen + closeMobilePanel（D12 reset）。"""
-        js = self._base_js()
-        assert "matchMedia" in js, "state-base.js 缺 matchMedia（D12 breakpoint reset 未實作）"
-        assert "960" in js, "state-base.js 缺 960（matchMedia 960px breakpoint）"
-        assert "similarModeMobileOpen" in js, \
-            "state-base.js 缺 similarModeMobileOpen（D12 reset 條件缺失）"
-        assert "closeMobilePanel" in js, \
-            "state-base.js 缺 closeMobilePanel 呼叫（D12 reset 動作缺失）"
-
-    # ── 12. burst-picker back.out exists, no 1.7 pinned ─────────────────────
-
-    def test_burst_picker_back_out_exists_no_1_7_pinned(self):
-        """burst-picker.js 含 back.out；且 guard 自身不 pin 具體 1.7 值（CD-11：doc/code drift）。"""
-        js = _T2_BURST_PICKER_JS.read_text(encoding="utf-8")
-        assert "back.out" in js, \
-            "burst-picker.js 缺 back.out（爆射 overshoot ease 未使用）"
-        # guard 自身不 assert 1.7（CD-11：不把 1.7 寫進守衛，避免文件/code drift）
-        # 只守 back.out 存在；具體值由 arcOvershoot param 決定，不硬鎖
-        # 反向：確保 burst-picker.js 仍用動態 arcOvershoot（非硬編碼 1.7 字串）
-        assert "arcOvershoot" in js, \
-            "burst-picker.js 缺 arcOvershoot param（back.out 值應來自 param，非硬編碼）"
-
-    # ── 13. Keydown intercept ────────────────────────────────────────────────
-
-    def test_mobile_panel_keydown_intercept(self):
-        """state-lightbox.js handleKeydown 函式體含 similarModeMobileOpen 條件分支 + closeMobilePanel。
-        且 similarModeMobileOpen guard block 不含 closeLightbox / prevLightboxVideo / nextLightboxVideo
-        （CD-7/D10：Esc 不關 lightbox，箭頭不切片）。
-        """
-        js = self._lightbox_js()
-        body = self._extract_function_body(js, r'handleKeydown\s*\(')
-        assert body is not None, \
-            "state-lightbox.js 找不到 handleKeydown 函式宣告"
-        assert "similarModeMobileOpen" in body, \
-            "handleKeydown 函式體缺 similarModeMobileOpen 條件（面板開時未攔截鍵盤）"
-        assert "closeMobilePanel" in body, \
-            "handleKeydown 函式體缺 closeMobilePanel 呼叫（Esc 未關行動面板）"
-        # 找 similarModeMobileOpen guard block，確認 block 不含 closeLightbox/prev/next
-        m = re.search(
-            r'if\s*\(\s*this\.similarModeMobileOpen\s*\)\s*\{([^}]*(?:\{[^}]*\}[^}]*)*)\}',
-            body, re.DOTALL
-        )
-        assert m, "handleKeydown 缺 if (this.similarModeMobileOpen) { ... } block"
-        guard_block = m.group(1)
-        assert "closeLightbox" not in guard_block, (
-            "handleKeydown similarModeMobileOpen block 含 closeLightbox()（Esc 不得關 lightbox，CD-7）"
-        )
-        assert "prevLightboxVideo" not in guard_block, (
-            "handleKeydown similarModeMobileOpen block 含 prevLightboxVideo（箭頭不得切片，D10）"
-        )
-        assert "nextLightboxVideo" not in guard_block, (
-            "handleKeydown similarModeMobileOpen block 含 nextLightboxVideo（箭頭不得切片，D10）"
-        )
 
 
 # ============================================================================
@@ -11679,212 +8995,6 @@ _T3_GHOST_FLY_JS = (
 )
 
 
-class TestMobilePanelT3Guards:
-    """TASK-83b-T3: 行動相似面板封面飛行進退場（6 條靜態守衛）。
-
-    每條 guard 均 mutation-detectable：刪除或改動對應實作即 RED。
-    """
-
-    @staticmethod
-    def _ghost_fly_js():
-        return _T3_GHOST_FLY_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _similar_js():
-        return _T2_SIMILAR_JS.read_text(encoding="utf-8")
-
-    @staticmethod
-    def _extract_function_body(js, func_pattern):
-        """大括號平衡法擷取函式體（reuse T2 pattern）。"""
-        m = re.search(func_pattern, js)
-        if not m:
-            return None
-        start = m.start()
-        try:
-            body_start = js.index('{', start)
-        except ValueError:
-            return None
-        depth = 0
-        for i, ch in enumerate(js[body_start:], body_start):
-            if ch == '{':
-                depth += 1
-            elif ch == '}':
-                depth -= 1
-                if depth == 0:
-                    return js[body_start:i + 1]
-        return None
-
-    # ── 1. Helper functions exported ────────────────────────────────────────
-
-    def test_mobile_panel_enter_exit_functions_exported(self):
-        """ghost-fly.js 含 playMobilePanelEnter + playMobilePanelExit 函式定義，
-        且兩者均出現在 GhostFly export 物件中。"""
-        js = self._ghost_fly_js()
-        assert "playMobilePanelEnter" in js, \
-            "ghost-fly.js 缺 playMobilePanelEnter 函式（83b-T3 helper 未建立）"
-        assert "playMobilePanelExit" in js, \
-            "ghost-fly.js 缺 playMobilePanelExit 函式（83b-T3 helper 未建立）"
-        # 確認 export 在 GhostFly 物件中（兩者均以 property: function 形式 export）
-        assert re.search(r'playMobilePanelEnter\s*:\s*playMobilePanelEnter', js), \
-            "ghost-fly.js GhostFly export 物件缺 playMobilePanelEnter 屬性"
-        assert re.search(r'playMobilePanelExit\s*:\s*playMobilePanelExit', js), \
-            "ghost-fly.js GhostFly export 物件缺 playMobilePanelExit 屬性"
-
-    # ── 2. Enter uses createCoverGhost, not constellation wrapper ───────────
-
-    def test_mobile_enter_uses_create_cover_ghost_not_constellation(self):
-        """ghost-fly.js playMobilePanelEnter 函式體含 createCoverGhost，
-        且不含 play56cConstellationEnter（不包裝桌面禁區函式）。
-        comment 中的字串不算（去掉 // 和 /* */ comment 後才驗證）。"""
-        js = self._ghost_fly_js()
-        body = self._extract_function_body(js, r'function\s+playMobilePanelEnter\s*\(')
-        assert body is not None, \
-            "ghost-fly.js 找不到 playMobilePanelEnter 函式宣告"
-        assert "createCoverGhost" in body, \
-            "playMobilePanelEnter 函式體缺 createCoverGhost 呼叫（必須直接用 primitive，不可包裝桌面函式）"
-        # 去掉 comment 後再搜尋（comment 中可能有說明性文字提及禁區函式名）
-        body_no_comments = re.sub(r'//[^\n]*', '', body)
-        body_no_comments = re.sub(r'/\*.*?\*/', '', body_no_comments, flags=re.DOTALL)
-        assert "play56cConstellationEnter" not in body_no_comments, \
-            "playMobilePanelEnter 函式體（去注釋後）含 play56cConstellationEnter（禁區：不可包裝桌面函式）"
-
-    # ── 3. Transition tokenized（DURATION.medium + fluent ease strings）──────
-
-    def test_mobile_transition_tokenized(self):
-        """ghost-fly.js playMobilePanelEnter / playMobilePanelExit 均：
-        - 含 DURATION.medium（token 不硬編碼）
-        - 含 0.333（fallback 允許）
-        - enter 含 fluent-decel；exit 含 fluent-accel
-        - 函式體（排除 fallback 表達式後）不含裸 duration 數字 literal（如 duration: 0.333）
-
-        Correction B：0.333 只允許作為 || fallback，不得作為 duration: 的直接值。
-        """
-        js = self._ghost_fly_js()
-
-        enter_body = self._extract_function_body(js, r'function\s+playMobilePanelEnter\s*\(')
-        assert enter_body is not None, "ghost-fly.js 找不到 playMobilePanelEnter"
-        exit_body = self._extract_function_body(js, r'function\s+playMobilePanelExit\s*\(')
-        assert exit_body is not None, "ghost-fly.js 找不到 playMobilePanelExit"
-
-        for name, body in [("playMobilePanelEnter", enter_body), ("playMobilePanelExit", exit_body)]:
-            assert "DURATION.medium" in body, \
-                f"{name} 函式體缺 DURATION.medium（必須用 token，不可硬編碼 duration）"
-            assert "0.333" in body, \
-                f"{name} 函式體缺 0.333 fallback（token guard chain 要求 || 0.333 fallback）"
-
-        assert "fluent-decel" in enter_body, \
-            "playMobilePanelEnter 函式體缺 fluent-decel ease（禁用 power2.* 代換）"
-        assert "fluent-accel" in exit_body, \
-            "playMobilePanelExit 函式體缺 fluent-accel ease（禁用 power2.* 代換）"
-
-        # Correction B：排除 || 0.333 fallback 行後，不得有 duration: <數字> 裸值
-        # 方法：把 "|| 0.333" fallback 表達式置換掉，再搜尋 duration: 後接數字
-        for name, body in [("playMobilePanelEnter", enter_body), ("playMobilePanelExit", exit_body)]:
-            body_stripped = body.replace("|| 0.333", "").replace("|| 0.5", "")
-            # 搜尋 duration: 後緊接數字 literal（含小數）
-            assert not re.search(r'\bduration\s*:\s*\d+(\.\d+)?', body_stripped), \
-                (f"{name} 函式體含裸 duration 數字 literal（排除 fallback 後仍殘留）；"
-                 "必須用 dur 變數（來自 DURATION.medium || 0.333）")
-
-    # ── 4. closeMobilePanel is async ────────────────────────────────────────
-
-    def test_mobile_close_panel_is_async(self):
-        """state-similar.js closeMobilePanel 為 async 函式（exit ghost await 前提）。"""
-        js = self._similar_js()
-        assert re.search(r'async\s+closeMobilePanel\s*\(', js), \
-            "state-similar.js closeMobilePanel 不是 async 函式（83b-T3：exit ghost 需 async + await）"
-
-    # ── 5. PRM fallback branches exist ──────────────────────────────────────
-
-    def test_mobile_transition_prm_fallback(self):
-        """state-similar.js _openMobilePanel 函式體含 shouldSkip（PRM 閘），
-        且 enter ghost 飛行（playMobilePanelEnter）被 shouldSkip 閘控（!...shouldSkip 在呼叫前）；
-        closeMobilePanel 含 shouldSkip + playMobilePanelExit 被閘控。"""
-        js = self._similar_js()
-
-        open_body = self._extract_function_body(js, r'async\s+_openMobilePanel\s*\(')
-        assert open_body is not None, \
-            "state-similar.js 找不到 _openMobilePanel 函式宣告"
-        assert "shouldSkip" in open_body, \
-            "_openMobilePanel 函式體缺 shouldSkip（PRM 降級分支未實作）"
-        assert "mobilePanelCoverImg" in open_body, \
-            "_openMobilePanel 函式體缺 mobilePanelCoverImg（PRM 分支中央主圖 src 設定缺失）"
-        # Nit 強化：enter ghost 飛行必須被 PRM 閘控（!...shouldSkip(...) 出現在 playMobilePanelEnter 呼叫之前）
-        assert "playMobilePanelEnter" in open_body, \
-            "_openMobilePanel 函式體缺 playMobilePanelEnter 呼叫（enter 飛行未接線）"
-        open_no_comments = re.sub(r'//[^\n]*', '', open_body)
-        open_no_comments = re.sub(r'/\*.*?\*/', '', open_no_comments, flags=re.DOTALL)
-        m_neg = re.search(r'!\s*window\.BurstPicker\.shouldSkip', open_no_comments)
-        m_enter = re.search(r'playMobilePanelEnter', open_no_comments)
-        assert m_neg is not None, \
-            "_openMobilePanel 缺 !window.BurstPicker.shouldSkip 否定閘（enter 飛行未受 PRM 閘控）"
-        assert m_neg.start() < m_enter.start(), (
-            "_openMobilePanel 的 !shouldSkip 閘必須在 playMobilePanelEnter 呼叫之前"
-            "（否則 PRM 用戶仍會跑進場飛行）"
-        )
-
-        close_body = self._extract_function_body(js, r'async\s+closeMobilePanel\s*\(')
-        assert close_body is not None, \
-            "state-similar.js 找不到 async closeMobilePanel 函式宣告"
-        assert "shouldSkip" in close_body, \
-            "closeMobilePanel 函式體缺 shouldSkip（PRM 快捷隱藏路徑未實作）"
-        # exit ghost 飛行同樣被 PRM 閘控
-        assert "playMobilePanelExit" in close_body, \
-            "closeMobilePanel 函式體缺 playMobilePanelExit 呼叫（exit 飛行未接線）"
-        close_no_comments = re.sub(r'//[^\n]*', '', close_body)
-        close_no_comments = re.sub(r'/\*.*?\*/', '', close_no_comments, flags=re.DOTALL)
-        m_neg_c = re.search(r'!\s*window\.BurstPicker\.shouldSkip', close_no_comments)
-        m_exit = re.search(r'playMobilePanelExit', close_no_comments)
-        assert m_neg_c is not None and m_neg_c.start() < m_exit.start(), (
-            "closeMobilePanel 的 !shouldSkip 閘必須在 playMobilePanelExit 呼叫之前"
-            "（否則 PRM 用戶仍會跑退場飛行）"
-        )
-
-    # ── 5b. closeMobilePanel kills in-flight enter timeline (P1-T3fix) ───────
-
-    def test_mobile_close_kills_enter_timeline(self):
-        """state-similar.js closeMobilePanel 函式體在 playMobilePanelExit 之前 kill in-flight
-        enter timeline（_mobileEnterTl.kill()）+ 顯式 cleanup enter ghost（_mobileEnterGhost）。
-        P1-T3fix：防中途打斷時 enter onComplete 在 exit 飛行中誤還原 mobileCoverEl opacity → 雙圖。"""
-        js = self._similar_js()
-        close_body = self._extract_function_body(js, r'async\s+closeMobilePanel\s*\(')
-        assert close_body is not None, \
-            "state-similar.js 找不到 async closeMobilePanel 函式宣告"
-        close_no_comments = re.sub(r'//[^\n]*', '', close_body)
-        close_no_comments = re.sub(r'/\*.*?\*/', '', close_no_comments, flags=re.DOTALL)
-        # 必須 kill in-flight enter timeline
-        m_kill = re.search(r'_mobileEnterTl\s*\.\s*kill\s*\(', close_no_comments)
-        assert m_kill is not None, (
-            "closeMobilePanel 缺 _mobileEnterTl.kill()（P1-T3fix：未 kill in-flight 進場 timeline → "
-            "中途打斷雙圖）"
-        )
-        # 必須顯式 cleanup enter ghost（.kill() 不保證 fire onInterrupt，故不可只靠 timeline 自清）
-        assert "_mobileEnterGhost" in close_no_comments, (
-            "closeMobilePanel 缺 _mobileEnterGhost 顯式 cleanup（.kill() 不 fire onInterrupt → "
-            "enter ghost 殘留 + opacity 未還原）"
-        )
-        # kill 必須在 exit 飛行（playMobilePanelExit）之前
-        m_exit = re.search(r'playMobilePanelExit', close_no_comments)
-        assert m_exit is not None, "closeMobilePanel 缺 playMobilePanelExit"
-        assert m_kill.start() < m_exit.start(), (
-            "_mobileEnterTl.kill() 必須在 playMobilePanelExit 之前"
-            "（否則 enter ghost 仍在飛 → 與 exit ghost 雙圖）"
-        )
-
-    # ── 6. Desktop constellation anchor not in mobile enter ─────────────────
-
-    def test_desktop_constellation_byte_identical_anchor(self):
-        """ghost-fly.js 含 .similar-main-anchor（桌面 anchor lookup 未被移除）；
-        且 playMobilePanelEnter 函式體不含 .similar-main-anchor（禁區 anchor 不被 mobile helper 引用）。"""
-        js = self._ghost_fly_js()
-        assert ".similar-main-anchor" in js, \
-            "ghost-fly.js 缺 .similar-main-anchor（桌面 play56cConstellationEnter 被破壞或移除）"
-
-        enter_body = self._extract_function_body(js, r'function\s+playMobilePanelEnter\s*\(')
-        assert enter_body is not None, \
-            "ghost-fly.js 找不到 playMobilePanelEnter 函式宣告"
-        assert ".similar-main-anchor" not in enter_body, \
-            "playMobilePanelEnter 函式體含 .similar-main-anchor（禁區：桌面 DOM 不得出現在 mobile helper）"
 
 
 class TestSimilarMobilePanelT4Guard:
@@ -11968,82 +9078,6 @@ class TestSimilarMobilePanelT4Guard:
             "ghost-hide 規則須含 pointer-events: none（飛行期間防誤點）"
 
 
-class TestHelpUpdateButtonGuard:
-    """84-T3: Help 頁「更新」按鈕 + confirm modal 靜態守衛
-
-    契約：
-    1. {% if is_desktop %} gate 存在於 help.html（按鈕被正確 gate）
-    2. 「更新」按鈕的 @click="triggerUpdate()" 存在於 gate 內
-    3. triggerUpdate() DOM binding 不應出現在 gate 外
-    4. help.js 定義 triggerUpdate function
-    5. modal 的 x-show="showUpdateModal" binding 存在
-    6. modal 內含 confirm / cancel 按鈕
-    """
-
-    HELP_HTML = PROJECT_ROOT / "web" / "templates" / "help.html"
-    HELP_JS   = PROJECT_ROOT / "web" / "static" / "js" / "pages" / "help.js"
-
-    def _html(self):
-        return self.HELP_HTML.read_text(encoding="utf-8")
-
-    def _js(self):
-        return self.HELP_JS.read_text(encoding="utf-8")
-
-    def test_is_desktop_gate_exists_in_help_html(self):
-        """help.html 含 {% if is_desktop %} gate（確保按鈕被正確 gate）"""
-        html = self._html()
-        assert '{% if is_desktop %}' in html, \
-            "help.html 缺 {% if is_desktop %} gate — 更新按鈕必須在此 gate 內"
-
-    def test_trigger_update_click_inside_desktop_gate(self):
-        """@click="triggerUpdate()" 必須在 {% if is_desktop %} block 內"""
-        import re
-        html = self._html()
-        gate_pattern = re.compile(
-            r'\{%-?\s*if\s+is_desktop\s*-?%\}(.*?)\{%-?\s*endif\s*-?%\}',
-            re.DOTALL,
-        )
-        gates = gate_pattern.findall(html)
-        assert gates, "help.html: 找不到 {% if is_desktop %} block"
-        found = any('triggerUpdate()' in block for block in gates)
-        assert found, (
-            '@click="triggerUpdate()" 必須在 {% if is_desktop %} block 內 '
-            '（非桌面客戶端不應看到此按鈕）'
-        )
-
-    def test_trigger_update_not_outside_gate(self):
-        """triggerUpdate() DOM binding 不應出現在 {% if is_desktop %} gate 外"""
-        import re
-        html = self._html()
-        gate_pattern = re.compile(
-            r'\{%-?\s*if\s+is_desktop\s*-?%\}.*?\{%-?\s*endif\s*-?%\}',
-            re.DOTALL,
-        )
-        stripped = gate_pattern.sub('', html)
-        assert 'triggerUpdate()' not in stripped, (
-            'triggerUpdate() 出現在 {% if is_desktop %} gate 外 — '
-            '非桌面情境不應渲染此 DOM'
-        )
-
-    def test_trigger_update_defined_in_help_js(self):
-        """help.js 定義 triggerUpdate function（防函數遺漏導致 Alpine 報錯）"""
-        js = self._js()
-        assert 'triggerUpdate' in js, \
-            "help.js 缺 triggerUpdate 定義 — Alpine @click 呼叫的函數不存在"
-
-    def test_update_modal_x_show_binding_exists(self):
-        """modal 含 x-show="showUpdateModal" binding（防按鈕存在但 modal 永不出現）"""
-        html = self._html()
-        assert 'showUpdateModal' in html, \
-            "help.html 缺 showUpdateModal binding — update confirm modal 無法顯示"
-
-    def test_update_modal_has_confirm_and_cancel(self):
-        """modal 內含 confirmUpdate() 和 cancelUpdate() 呼叫（confirm/cancel 按鈕齊全）"""
-        html = self._html()
-        assert 'confirmUpdate()' in html, \
-            "help.html 缺 confirmUpdate() — modal 確認按鈕不存在"
-        assert 'cancelUpdate()' in html, \
-            "help.html 缺 cancelUpdate() — modal 取消按鈕不存在"
 
 
 class TestDirPathHelperGuard:
